% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_velocity.R
\name{acc_velocity_model}
\alias{acc_velocity_model}
\title{Instantaneous Velocity During Running Acceleration}
\usage{
acc_velocity_model(time, tau, maximal_velocity, reaction_time = 0)
}
\arguments{
\item{time}{A vector of time points (in s)}

\item{tau}{A double representing the time constant of the exponential rise in velocity (in s)}

\item{maximal_velocity}{A double representing the maximal velocity (in m/s)}

\item{reaction_time}{A double representing the reaction time (in s) Default value is 0}
}
\value{
A vector of instantaneous velocities (in m/s)
}
\description{
Estimates the instantaneous velocity during the acceleration phase of a sprint run. If the reaction time is known,
it can be included in the model. The model is based on an exponential rise in velocity.
}
\examples{
acc_velocity_model(1:10, 0.5, 10)
acc_velocity_model(1:10, 0.5, 10, reaction_time = 0.1)
}
