% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_velocity.R
\name{acceleration_sprint_model}
\alias{acceleration_sprint_model}
\title{Instantaneous Acceleration During Sprint Running (Acceleration and Deceleration phases)}
\usage{
acceleration_sprint_model(
  time,
  time_maximal_velocity,
  fitted_maximal_velocity,
  tau,
  decel_rate
)
}
\arguments{
\item{time}{A vector of time points (in s)}

\item{time_maximal_velocity}{A double representing the time at which maximal velocity is reached (in s)}

\item{fitted_maximal_velocity}{A double representing the maximal velocity fitted using the model for instantaneous velocity in acceleration phase (in m/s)}

\item{tau}{A double representing the time constant of the exponential rise in velocity (in s)}

\item{decel_rate}{A double representing the rate of deceleration (in m/s^2)}
}
\value{
A vector of instantaneous accelerations (in m/s^2)
}
\description{
Estimates the instantaneous acceleration during the acceleration and deceleration phases of a sprint run.
The calculation is based on an exponential rise in velocity during the acceleration phase and a constant rate of deceleration during the deceleration phase.
}
\examples{
acceleration_sprint_model(1:10, 5, 10, 1.5, 0.1)

}
