% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_integrals.R
\name{aerobic_energy_percentage}
\alias{aerobic_energy_percentage}
\title{Aerobic Energy Percentage}
\usage{
aerobic_energy_percentage(data, type = "power")
}
\arguments{
\item{data}{A tibble with a time series of aerobic power values and associated times. The tibble must minimally have the following columns: time (s), power (W/kg), power_aerobic (W/kg)}

\item{type}{A string indicating the type of power to integrate. Default is "power" and refers to estimated metabolic power derived from di Prampero et al. equivalent slope approach. "power model" refers to the power estimated using the sprint bioenergetic model.}
}
\value{
A double representing the percentage (\%) of anaerobic aerobic energy expenditure relative to total energy expenditure.
}
\description{
Computes the percentage of the total energy expenditure that is attributed to aerobic energy pathway.
}
\examples{

data <- tibble::tibble(
  time = c(0, 1, 2, 3, 4),
  power = c(0, 1, 2, 1, 0),
  power_aerobic = c(0, 0.5, 1, 0.5, 0)
)

aerobic_energy_percentage(data)

}
