% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_motion_computations.R
\name{average_velocity_from_splits}
\alias{average_velocity_from_splits}
\title{Average Velocity From Splits}
\usage{
average_velocity_from_splits(splits, distances)
}
\arguments{
\item{splits}{a vector of time splits (in s)}

\item{distances}{a vector of distances (in m)}
}
\value{
a vector of average velocities (in m/s)
}
\description{
Computes average velocity over time intervals based on recorded time splits and their
associated distances.
}
\examples{
average_velocity_from_splits(c(0, 1.8), c(0, 10))
}
