% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bi_exponential_knorm}
\alias{bi_exponential_knorm}
\title{Bi-exponential Function Normalisation Constant}
\usage{
bi_exponential_knorm(t1, t2)
}
\arguments{
\item{t1}{a double representing the time constant of the rising exponential term (in s)}

\item{t2}{a double representing the time constant of the decaying exponential term (in s)}
}
\value{
a double representing the normalization constant
}
\description{
Calculates a normalization constant using the time constants of the rising and decaying exponential terms,
ensuring that the amplitude term preceding the exponential functions matches the maximum value achieved by the bi-exponential function.
}
\examples{

bi_exponential_knorm(2, 40)

}
\keyword{internal}
