% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolic_conversions.R
\name{convert_acc_lactate_to_lactic_energy}
\alias{convert_acc_lactate_to_lactic_energy}
\title{Accumulated Lactate to Lactic Energy Conversion}
\usage{
convert_acc_lactate_to_lactic_energy(acc_lactate, sex = "male")
}
\arguments{
\item{acc_lactate}{A numeric value representing accumulated lactate in mmol/L.}

\item{sex}{A character corresponding to the athlete sex (default is \code{"male"}, other option is \code{"female"}).}
}
\value{
A numeric value representing lactic energy in J/kg.
}
\description{
Converts accumulated lactate (mmol/L) to lactic energy (J/kg), considering the accumulation
of 1 mmol L-1 of blood lactate corresponds to an energy expenditure from the
anaerobic lactic metabolism equivalent to the consumption of 3 mL O2/kg of body mass for males and
2.7 mL O2/kg of body mass for females and considering an O2 energy equivalent of 20.9 J/mL. For more details,
see Briand et al. (2025) and di Prampero et al. (1981).
}
\examples{

acc_lactate <- 10  # Example accumulated lactate in mmol/L
convert_acc_lactate_to_lactic_energy(acc_lactate)


}
