% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolic_conversions.R
\name{convert_aerobic_power_to_vo2}
\alias{convert_aerobic_power_to_vo2}
\title{Aerobic Power to VO2 Conversion}
\usage{
convert_aerobic_power_to_vo2(aerobic_power)
}
\arguments{
\item{aerobic_power}{A numeric value representing aerobic power in W/kg.}
}
\value{
A numeric value representing VO2 in mL/kg/min.
}
\description{
Converts aerobic power (W/kg) to VO2 (mL/kg/min), considering the energy equivalent
of oxygen of 20.9 mL O2/kg/min. For more details, see Briand et al. (2025) and di Prampero (1981).
}
\examples{
aerobic_power <- 20  # Example aerobic power in W/kg
convert_aerobic_power_to_vo2(aerobic_power)


}
