% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolic_conversions.R
\name{convert_lactic_energy_to_acc_lactate}
\alias{convert_lactic_energy_to_acc_lactate}
\title{Lactic Energy to Accumulated Lactate Conversion}
\usage{
convert_lactic_energy_to_acc_lactate(lactic_energy, sex = "male")
}
\arguments{
\item{lactic_energy}{A numeric value representing lactic energy in J/kg.}

\item{sex}{A character corresponding to the athlete sex (default is \code{"male"}, other option is \code{"female"}).}
}
\value{
A numeric value representing accumulated lactate in mmol/L.
}
\description{
Converts lactic energy (J/kg) to accumulated lactate (mmol/L), considering the accumulation
of 1 mmol L-1 of blood lactate corresponds to an energy expenditure from the
anaerobic lactic metabolism equivalent to the consumption of 3 mL O2/kg of body mass for males and
2.7 mL O2/kg of body mass for females and considering an O2 energy equivalent of 20.9 J/mL. For more details,
see Briand et al. (2025) and di Prampero et al. (1981).
}
\examples{

lactic_energy <- 1000  # Example lactic energy in J/kg
convert_lactic_energy_to_acc_lactate(lactic_energy)


}
