% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolic_conversions.R
\name{convert_vo2_to_aerobic_power}
\alias{convert_vo2_to_aerobic_power}
\title{VO2 to Aerobic Power Conversion}
\usage{
convert_vo2_to_aerobic_power(vo2)
}
\arguments{
\item{vo2}{A numeric value representing VO2 in mL/kg/min.}
}
\value{
A numeric value representing aerobic power in W/kg.
}
\description{
Converts VO2 (mL/kg/min) to aerobic power (W/kg), considering the energy equivalent
of oxygen of 20.9 mL O2/kg/min. For more details, see Briand et al. (2025) and di Prampero (1981).
}
\examples{
vo2 <- 50  # Example VO2 in mL/kg/min
convert_vo2_to_aerobic_power(vo2)


}
