% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_of_running.R
\name{cost_running_sprint}
\alias{cost_running_sprint}
\title{Cost of Running (Sprint Version)}
\usage{
cost_running_sprint(acceleration, velocity)
}
\arguments{
\item{acceleration}{a vector of accelerations (m/s^2)}

\item{velocity}{a vector of velocities (m/s)}
}
\value{
a vector of costs of running (J/kg/m)
}
\description{
Computes the instantaneous cost of running based on velocity and acceleration data,
based on the method of di prampero et al. (2005, 2018) and as described in Briand et al. (2025).
In the sprint application, it is assumed that acceleration cannot be negative.
The observed slowdown in sprints is not due to a change in the runner's
inclination of the center of mass but rather the inability to generate the required power.
}
\examples{
cost_running_sprint(acceleration = c(2, 4, 5), velocity = c(2, 4, 5))

}
