% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_motion_computations.R
\name{dec_velocity_model}
\alias{dec_velocity_model}
\title{Instantaneous Velocity During Sprint Running Deceleration Phase}
\usage{
dec_velocity_model(time, maximal_velocity, time_maximal_velocity, decel_rate)
}
\arguments{
\item{time}{A vector of time points (in s)}

\item{maximal_velocity}{A double representing the maximal velocity (in m/s)}

\item{time_maximal_velocity}{A double representing the time at which maximal velocity is reached (in s)}

\item{decel_rate}{A double representing the rate of deceleration (in m/s^2)}
}
\value{
A vector of instantaneous velocities (in m/s)
}
\description{
Estimates the instantaneous velocity during the deceleration phase of a sprint run. The model is based on a constant rate of deceleration.
}
\examples{
dec_velocity_model(6:10, 10, 6, 0.1)
}
