% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_motion_models_diagnostics.R
\name{diagnose_sprint_model}
\alias{diagnose_sprint_model}
\title{Test the Assumptions of the Sprint Motion Model Fit}
\usage{
diagnose_sprint_model(
  mean_velocity_splits,
  time_splits,
  distance,
  reaction_time,
  maximal_velocity = NA,
  dt = 0.01,
  cost_running_flat = 3.8,
  slope_equation = "original"
)
}
\arguments{
\item{mean_velocity_splits}{A vector with mean velocity splits over each distance interval (m/s)}

\item{time_splits}{A vector with time splits over each distance interval (s)}

\item{distance}{A vector with the distances at which time splits were measured (m)}

\item{reaction_time}{A double with the reaction time measured on the starting blocks (s)}

\item{maximal_velocity}{A double representing the maximal velocity (in m/s)}

\item{dt}{Time step of the model. Default is set at 0.01 s}

\item{cost_running_flat}{a numeric value representing the cost of running on a flat surface (default for this function is 3.8 J/kg/m, as used in Briand et al. 2025)}

\item{slope_equation}{a string indicating the slope equation to use, either "original" or "extended". The "original" equation is based on the equation relating running at constant speed on different slopes porposed by Minetti et al. (2002), while the "extended" equation is based on similar equation for extended slopes proposed by Minetti and Pavei (2018).}
}
\value{
A list containing the RMSE, RSE, a tibble with observed vs fitted velocities and a list of ggplot objects for each diagnostic plot.
}
\description{
Test the Assumptions of the Sprint Motion Model Fit
}
\examples{

men_200 <- graubner_nixdorf_sprints |>
  dplyr::filter(event == "Men's 200 m")

diagnose_sprint_model(
  mean_velocity_splits = men_200$velocity,
  time_splits = men_200$splits,
  distance = men_200$distance,
  reaction_time = men_200$reaction_time[1],
  maximal_velocity = men_200$maximal_velocity[1]
)

}
