% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_velocity.R
\name{distance_sprint_model}
\alias{distance_sprint_model}
\title{Instantaneous Distance During Sprint Running (Acceleration and Deceleration phases)}
\usage{
distance_sprint_model(
  time,
  time_maximal_velocity,
  maximal_velocity,
  distance_maximal_velocity,
  fitted_maximal_velocity,
  tau,
  decel_rate
)
}
\arguments{
\item{time}{A vector of time points (in s)}

\item{time_maximal_velocity}{A double representing the time at which maximal velocity is reached (in s)}

\item{maximal_velocity}{A double representing the maximal velocity (in m/s)}

\item{distance_maximal_velocity}{A double representing the distance at which maximal velocity is reached (in m)}

\item{fitted_maximal_velocity}{A double representing the maximal velocity fitted using the model for instantaneous velocity in acceleration phase (in m/s)}

\item{tau}{tau A double representing the time constant of the exponential rise in velocity (in s)}

\item{decel_rate}{A double representing the rate of deceleration (in m/s^2)}
}
\value{
A vector of instantaneous distances (in m)
}
\description{
Instantaneous Distance During Sprint Running (Acceleration and Deceleration phases)
}
\examples{
distance_sprint_model(1:10, 5, 10, 100, 10, 1.5, 0.1)

}
