% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_integrals.R
\name{energy_aerobic}
\alias{energy_aerobic}
\title{Aerobic Energy Expenditure}
\usage{
energy_aerobic(data)
}
\arguments{
\item{data}{A tibble with a time series of aerobic power values and associated times. The tibble must minimally have the following columns: time (s), power_aerobic (W/kg)}
}
\value{
A double corresponding to aerobic energy expended (J/kg)
}
\description{
Computes the integral of the aerobic power over time to obtain aerobic energy expenditure.
}
\examples{
data <- tibble::tibble(
  time = c(0, 1, 2, 3, 4),
  power_aerobic = c(0, 1, 2, 1, 0)
)

energy_aerobic(data)

}
