% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_integrals.R
\name{energy_alactic}
\alias{energy_alactic}
\title{Anaerobic alactic Energy Expenditure}
\usage{
energy_alactic(data)
}
\arguments{
\item{data}{A tibble with a time series of alactic power values and associated times. The tibble must minimally have the following columns: time (s), power_alactic (W/kg)}
}
\value{
A double corresponding to alactic energy expended (J/kg)
}
\description{
Computes the integral of the anaerobic alactic power over time to obtain anaerobic alactic energy expenditure.
}
\examples{
data <- tibble::tibble(
  time = c(0, 1, 2, 3, 4),
  power_alactic = c(0, 1, 2, 1, 0)
)

energy_alactic(data)

}
