% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_integrals.R
\name{energy_anaerobic}
\alias{energy_anaerobic}
\title{Anaerobic Energy Expenditure}
\usage{
energy_anaerobic(data)
}
\arguments{
\item{data}{A tibble with a time series of anaerobic power values and associated times. The tibble must minimally have the following columns: time (s), power_anaerobic (W/kg)}
}
\value{
A double corresponding to anaerobic energy expended (J/kg)
}
\description{
Computes the integral of the anaerobic power over time to obtain anaerobic energy expenditure.
}
\examples{
data <- tibble::tibble(
  time = c(0, 1, 2, 3, 4),
  power_anaerobic = c(0, 1, 2, 1, 0)
)

energy_anaerobic(data)

}
