% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_integrals.R
\name{energy_lactic}
\alias{energy_lactic}
\title{Anaerobic Lactic Energy Expenditure}
\usage{
energy_lactic(data)
}
\arguments{
\item{data}{A tibble with a time series of lactic power values and associated times. The tibble must minimally have the following columns: time (s), power_lactic (W/kg)}
}
\value{
A double corresponding to lactic energy expended (J/kg)
}
\description{
Computes the integral of the anaerobic lactic power over time to obtain anaerobic lactic energy expenditure.
}
\examples{
data <- tibble::tibble(
  time = c(0, 1, 2, 3, 4),
  power_lactic = c(0, 1, 2, 1, 0)
)

energy_lactic(data)

}
