% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_integrals.R
\name{energy_total}
\alias{energy_total}
\title{Total Energy Expenditure}
\usage{
energy_total(data, type = "power")
}
\arguments{
\item{data}{A tibble with a time series of power values and associated times. The tibble must minimally have the following columns: time (s), power (W/kg)}

\item{type}{A string indicating the type of power to integrate. Default is "power" and refers to estimated metabolic power derived from di Prampero et al. equivalent slope approach. "power model" refers to the power estimated using the sprint bioenergetic model.}
}
\value{
A double with the total energy expended (J/kg)
}
\description{
Computes the integral of the total metabolic power over time to obtain total energy expenditure. The total energy expenditure estimated the sprint bioenergetic model
can also be computed by entering an extra column (power_mod) corresponding to the power derived from the sprint bioenergetic model.The \code{type} parameter has to be adjusted accordingly.
}
\examples{
data <- tibble::tibble(
  time = c(0, 1, 2, 3, 4),
  power = c(0, 1, 2, 1, 0)
)

energy_total(data)

}
