% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_power.R
\name{external_power}
\alias{external_power}
\title{External (mechanical) running power}
\usage{
external_power(acceleration, velocity)
}
\arguments{
\item{acceleration}{Numeric vector of accelerations (m/s^2).}

\item{velocity}{Numeric vector of velocities (m/s).}
}
\value{
A numeric vector of external mechanical power (W/kg).
}
\description{
Instantaneous external mechanical power per unit body mass from acceleration
and velocity, following di Prampero et al. (2024), \emph{Mechanical and Metabolic
Power in Accelerated Running - PART I: the 100-m dash}.
}
\examples{
external_power(acceleration = c(0, 2, 4), velocity = c(2, 4, 6))
}
