% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_lactic_energy_duration.R
\name{find_max_la}
\alias{find_max_la}
\title{Find Sprint Maximal Lactic Power}
\usage{
find_max_la(duration, lactic_capacity, t1 = 20, t2 = 1500, k1 = 2.75, k2 = 35)
}
\arguments{
\item{duration}{A numeric value representing the sprint duration (in seconds).}

\item{lactic_capacity}{A numeric value representing the lactic capacity (in J/kg).}

\item{t1}{A numeric value representing the time constant for the first exponential decay (default is 20 seconds).}

\item{t2}{A numeric value representing the time constant for the second exponential decay (default is 1500 seconds).}

\item{k1}{A double. Time constant of the first rising exponential (s). Default is 2.75}

\item{k2}{A double. Time constant of the second decaying exponential (s). Default is 35}
}
\value{
A numeric value representing the maximal lactic power (in W/kg) that can be sustained over the given sprint duration.
}
\description{
Knowing an athlete maximal lactic capacity, the function finds
the maximal lactic power that can be sustained over a given sprint duration
}
\examples{

duration <- 20
lactic_capacity <- 1350

find_max_la(duration, lactic_capacity)

}
\keyword{internal}
