% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_motion_computations.R
\name{find_tau}
\alias{find_tau}
\title{Finding Acceleration Exponential Rise Time-Constant (tau)}
\usage{
find_tau(time, velocity, reaction_time)
}
\arguments{
\item{time}{A vector of time points (in s)}

\item{velocity}{A vector of velocities (in s)}

\item{reaction_time}{A double representing the reaction time (in s)}
}
\value{
A double representing the time constant of the exponential rise in velocity (in s)
}
\description{
Computes the time constant of the exponential rise in velocity during the acceleration phase of a sprint run
using vectors of assumed instantaneous velocities and times throughout the sprint acceleration.
}
\examples{
find_tau(1:10, 1:10, 0.1)
}
