% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_velocity.R
\name{find_terminal_velocity}
\alias{find_terminal_velocity}
\title{Find Terminal Velocity}
\usage{
find_terminal_velocity(splits, velocity, distance, reaction_time)
}
\arguments{
\item{splits}{A vector containing time splits (in s)}

\item{velocity}{A vector containing average velocity over each interval (in m/s)}

\item{distance}{A vector containing intermediate distances at which splits are measured (in m)}

\item{reaction_time}{A double corresponding to the reaction time (in s)}
}
\value{
A double representing the terminal velocity (in m/s)
}
\description{
Estimates the velocity at the end of the race. Based on splits data measured at different intervals
and the average velocity over each interval.
}
\examples{
find_terminal_velocity(c(8.74, 9.58), c( 12.17, 11.96), c(90, 100), 0.146)

}
