% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_velocity.R
\name{find_time_velocity}
\alias{find_time_velocity}
\title{Find Times Associated With Velocity Measurements}
\usage{
find_time_velocity(splits, reaction_time)
}
\arguments{
\item{splits}{A vector containing time splits (in s)}

\item{reaction_time}{A double corresponding to the reaction time (in s)}
}
\value{
A vector of times (in s) where instantaneous velocity measurements occur
}
\description{
In the cases where data are provided as time splits, with reaction time and average velocity
for each time intervals, it can be assumed that the instantaneous velocity measurement occurs at the
middle point of each interval. The function returns a vector of times where instantaenous velocities occur.
}
\examples{
find_time_velocity( c(0, 1.88, 2.88, 3.78, 4.64, 5.47, 6.29, 7.10, 7.92, 8.74, 9.58), 0.173)

}
