% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_energetics_exploration.R
\name{fit_approx_lactic_gof_metrics}
\alias{fit_approx_lactic_gof_metrics}
\title{Goodness of Fit (Gof) Metrics for Approximate Lactic Power Model}
\usage{
fit_approx_lactic_gof_metrics(data)
}
\arguments{
\item{data}{A tibble containing approximate power distributions over the course of the sprint with at least the following columns: time (s), power_lactic (W/kg)}
}
\value{
A list with the following elements: r_squared, residual_se, AIC, BIC
}
\description{
Computes the goodness of fit metrics for the approximate lactic power model.
}
\examples{
# Extract the data for the 100 m
men_100 <- graubner_nixdorf_sprints |>
  dplyr::filter(event == "Men's 100 m")


# Get the sprint motion data for both men and women

sprint_data <- sprint_motion_model_data(
  mean_velocity_splits = men_100$velocity,
  time_splits = men_100$splits,
  distance = men_100$distance,
  reaction_time = men_100$reaction_time[1],
  maximal_velocity = men_100$maximal_velocity[1]
)

sprint_approx_power_distributions <- sprint_approx_power_distributions(sprint_data,
  maximal_aerobic_power = 24.5,
  basal_metabolic_rate = 1.2
)

# Fit the log-normal distribution to the alactic power distribution

fit_approx_lactic_gof_metrics(sprint_approx_power_distributions)

}
