% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_lactic_model_formalisation.R
\name{fit_lactic_power_params}
\alias{fit_lactic_power_params}
\title{Parameters of formalized/simplified Log-normal Fit on Lactic Power Distribution}
\usage{
fit_lactic_power_params(sprint_approx_power_distribution, k1 = 2.75, k2 = 35)
}
\arguments{
\item{sprint_approx_power_distribution}{A tibble containing approximate power distributions over the course of the sprint with at least the following columns: time (s), power_lactic (W/kg)}

\item{k1}{A double. Time constant of the first rising exponential (s).}

\item{k2}{A double. Time constant of the second decaying exponential (s).}
}
\value{
A list with the following elements: p_la_max, k1, k2
}
\description{
Extracts the parameters of a simplified/formalized bi-exponential fit on the approximate lactic power distribution. The formalized model is a model where the k1 and k2 parameters of the bi-exponential distribution are set at 2.75 and 35 s, respectively.
The only fitted value is therefore the maximal lactic power. The set value correspond to a compromise of the observed fitted values over the different sprint events.
}
\examples{
# Extract the data for the 100 m
men_100 <- graubner_nixdorf_sprints |>
  dplyr::filter(event == "Men's 100 m")


# Get the sprint motion data for both men and women

sprint_data <- sprint_motion_model_data(
  mean_velocity_splits = men_100$velocity,
  time_splits = men_100$splits,
  distance = men_100$distance,
  reaction_time = men_100$reaction_time[1],
  maximal_velocity = men_100$maximal_velocity[1]
)

sprint_approx_power_distributions <- sprint_approx_power_distributions(sprint_data,
  maximal_aerobic_power = 24.5,
  basal_metabolic_rate = 1.2
)

# Fit the log-normal distribution to the alactic power distribution

fit_lactic_power_params(sprint_approx_power_distributions)

}
