% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sprint_alactic_duration_model.R
\name{get_alactic_model_data}
\alias{get_alactic_model_data}
\title{Generate model predictions for the sprint alactic power-duration model}
\usage{
get_alactic_model_data(
  alactic_power_duration,
  sex_label,
  mu_al = 1.75,
  sigma_al = 1.5
)
}
\arguments{
\item{alactic_power_duration}{A tibble with the following columns: duration (s), alactic_power (W/kg), corresponding to the average alactic power output over the running duration. This data is used to fit the alactic power model.}

\item{sex_label}{A string label (e.g., "male", "Female") used for grouping the output.}

\item{mu_al}{A double representing the peak of the log-normal distribution. Default is 1.75.}

\item{sigma_al}{A double representing the decay of the log-normal distribution. Default is 1.5.}
}
\value{
A tibble with \code{duration}, \code{power}, and \code{sex} columns representing the predicted model values.
}
\description{
Computes the modeled alactic power output across a sequence of durations,
based on the estimated alactic capacity from empirical data and log-normal decay alactic power duration model.
}
\examples{
data <- sprint_alactic_energy_duration_graubner_nixdorf()
get_alactic_model_data(data, sex_label = "male")
}
