% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_perf_from_splits.R
\name{graubner_nixdorf_perf_from_splits}
\alias{graubner_nixdorf_perf_from_splits}
\title{Extract Sprint Performance for Graubner and Nixdorf (2011) Time Splits Data}
\usage{
graubner_nixdorf_perf_from_splits()
}
\value{
A tibble with the time performance, over each distance of the graubner_nixdorf_sprints dataset, with associated event and athlete sex.
}
\description{
Extracts the sprint performance (time to cover the total distance) for each event in the Graubner and Nixdorf (2011) dataset.
}
\examples{
#graubner_nixdorf_perf_from_splits()

}
