% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{graubner_nixdorf_sprints}
\alias{graubner_nixdorf_sprints}
\title{Graubner and Nixdorf sprint data}
\format{
\subsection{\code{graubner_nixdorf_sprints}}{

A tibble with 45 rows and 6 columns:
\describe{
\item{distance}{Distance (in m) at which time splits were measured for each event}
\item{splits}{Time splits (in s)}
\item{velocity}{Average velocity (m/s) over each distance interval}
\item{reaction_time}{Reaction time (s) measured on the starting blocks at the beginning of the race}
\item{maximal_velocity}{Maximal velocity (m/s) measured during the race. If not available, NA}
\item{athlete}{Name of the athlete}
\item{event}{Race event at the World athletics championships}
}
}
}
\source{
Graubner, R., & Nixdorf, E. (2009). Biomechanical analysis of the sprint and hurdles events at the 2009 IAAF World Championships in Athletics. Positions, 1(10).
}
\usage{
graubner_nixdorf_sprints
}
\description{
Time splits over distances recorded at the 2009 World athletics Championships in Berlin.
}
\keyword{datasets}
