% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kindermann_lactate}
\alias{kindermann_lactate}
\title{Kindermann lactate data}
\format{
\subsection{\code{kindermann_lactate}}{

A tibble with 10 rows and 3 columns:
\describe{
\item{duration}{Duration (in s) of the running event over which lactate was measured}
\item{lactate}{Blood lactate concentration (in mmol/L)}
\item{accumulated_lactate}{Accumulated lactate above blood lactate resting levels, which are assumed to be 1 mmol/L}
}
}
}
\source{
Kindermann, W. (1977). Lactate acidosis with different forms of sports activities. Can. J. Appl. Sports Sci. https://cir.nii.ac.jp/crid/1574231873976904704
}
\usage{
kindermann_lactate
}
\description{
Provides lactate and accumulated lactate values over running durations
}
\keyword{datasets}
