% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{middle_point}
\alias{middle_point}
\title{Middle Point}
\usage{
middle_point(t1, t2)
}
\arguments{
\item{t1}{a double representing the time of the first boundary of the interval (in s)}

\item{t2}{a double representing the time of the second boundary of the interval (in s)}
}
\value{
a double representing the midpoint time (in s)
}
\description{
Calculates the midpoint time within a given time interval.
}
\examples{
middle_point(2, 4)

}
\keyword{internal}
