% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_running_power_computations.R
\name{plot_metabolic_external_power_comparison}
\alias{plot_metabolic_external_power_comparison}
\title{Plot metabolic and external power with efficiency}
\usage{
plot_metabolic_external_power_comparison(
  power_data,
  color_metabolic = "darkblue",
  color_external = "darkred",
  color_efficiency = "darkgreen"
)
}
\arguments{
\item{power_data}{A data frame with \code{time}, \code{metabolic_power}, \code{external_power}
and \code{external_mechanical_efficiency} columns, e.g. from
\code{\link[=compute_external_mechanical_efficiency]{compute_external_mechanical_efficiency()}}.}

\item{color_metabolic}{Colour of the metabolic-power line.}

\item{color_external}{Colour of the external-power line.}

\item{color_efficiency}{Colour of the efficiency line.}
}
\value{
A patchwork of the power and efficiency plots.
}
\description{
Plot metabolic and external power with efficiency
}
\examples{
\donttest{
f <- system.file("extdata/activities/sample_run.fit", package = "runrgetics")
if (nzchar(f)) plot_metabolic_external_power_comparison(compute_activity_power_efficiency_data(f))
}
}
