% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_running_power_computations.R
\name{plot_running_analysis}
\alias{plot_running_analysis}
\title{Plot a full running analysis (power, motion, efficiency, heart rate)}
\usage{
plot_running_analysis(
  running_data,
  color_metabolic_power = "darkblue",
  color_external_power = "darkred",
  color_velocity = "purple",
  color_acceleration = "darkorange",
  color_heart_rate = "red",
  color_efficiency = "darkgreen"
)
}
\arguments{
\item{running_data}{A data frame with \code{time}, \code{metabolic_power}, \code{external_power},
\code{external_mechanical_efficiency}, \code{velocity}, \code{acceleration} and \code{heartrate}
columns.}

\item{color_metabolic_power}{Colour of the metabolic-power line.}

\item{color_external_power}{Colour of the external-power line.}

\item{color_velocity}{Colour of the velocity line.}

\item{color_acceleration}{Colour of the acceleration line.}

\item{color_heart_rate}{Colour of the heart-rate line.}

\item{color_efficiency}{Colour of the efficiency line.}
}
\value{
A patchwork of the power, motion, efficiency and heart-rate plots.
}
\description{
Plot a full running analysis (power, motion, efficiency, heart rate)
}
\examples{
\donttest{
f <- system.file("extdata/activities/sample_run.fit", package = "runrgetics")
if (nzchar(f)) plot_running_analysis(compute_activity_power_efficiency_data(f))
}
}
