% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sprint_motion_analysis.R
\name{plot_sprint_acceleration}
\alias{plot_sprint_acceleration}
\title{Plot sprint acceleration as a function of time}
\usage{
plot_sprint_acceleration(
  mean_velocity_splits,
  time_splits,
  distance,
  reaction_time,
  maximal_velocity = NA,
  dt = 0.01,
  color = "blue"
)
}
\arguments{
\item{mean_velocity_splits}{A vector with mean velocity splits over each distance interval (m/s)}

\item{time_splits}{A vector with time splits over each distance interval (s)}

\item{distance}{A vector with the distances at which time splits were measured (m)}

\item{reaction_time}{A double with the reaction time measured on the starting blocks (s)}

\item{maximal_velocity}{A double representing the maximal velocity (in m/s)}

\item{dt}{Time step of the model. Default is set at 0.01 s}

\item{color}{A string with the color of the plot's line and points. Default is set at "blue"}
}
\value{
A ggplot object
}
\description{
Plot of modeled sprint acceleration as a function of time.
}
\examples{
 plot_sprint_acceleration(
  mean_velocity_splits = c(0, 5.77, 9.99),
  time_splits = c(0, 1.88, 2.88),
  distance = c(0, 10, 20),
  reaction_time = 0.146,
  maximal_velocity = 12.34
)

}
