% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sprint_alactic_duration_model.R
\name{plot_sprint_alactic_duration}
\alias{plot_sprint_alactic_duration}
\title{Plot Sprint Alactic Power vs Duration with Model Overlay}
\usage{
plot_sprint_alactic_duration(
  alactic_power_duration,
  mu_al = 1.75,
  sigma_al = 1.5,
  linetype = "solid",
  line_color = "purple",
  point_color = "darkorchid4",
  point_shape = 16
)
}
\arguments{
\item{alactic_power_duration}{A tibble with the following columns: duration (s), alactic_power (W/kg), corresponding to the average alactic power output over the running duration. This data is used to fit the alactic power model.}

\item{mu_al}{A double representing the peak of the log-normal distribution. Default is 1.75.}

\item{sigma_al}{A double representing the decay of the log-normal distribution. Default is 1.5.}

\item{linetype}{Line type used for the model curve (default = \code{"solid"}). Accepts any valid ggplot2 line type.}

\item{line_color}{Color of the model prediction line (default = \code{"purple"}).}

\item{point_color}{Color of the observed data points (default = \code{"darkorchid4"}).}

\item{point_shape}{Shape of the observed data points (default = 16, i.e., solid circle).}
}
\value{
A \code{ggplot} object showing the observed alactic power-duration data and the fitted model curve.
}
\description{
Creates a ggplot visualizing observed alactic power output as a function of sprint duration,
overlaid with a model prediction curve based on fitted alactic power duration model.
}
\examples{
# Example with sample sprint data
data <- tibble::tibble(
  duration = c(2, 4, 6, 8),
  alactic_power = c(350, 320, 300, 290)
)
plot_sprint_alactic_duration(data)

# Other example with graubner_nixdorf_sprints data

data <- sprint_alactic_energy_duration_graubner_nixdorf()

plot_sprint_alactic_duration(data)

}
