% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sprint_alactic_duration_model.R
\name{plot_sprint_alactic_duration_briand_article}
\alias{plot_sprint_alactic_duration_briand_article}
\title{Plot alactic power-duration model fits for male and female sprinters}
\usage{
plot_sprint_alactic_duration_briand_article(
  data = graubner_nixdorf_sprints,
  mu_al = 1.75,
  sigma_al = 1.5,
  line_color = "purple",
  point_color = "darkorchid4",
  mu = -0.4,
  sigma = 1,
  k1 = 2.75,
  k2 = 35,
  dt = 0.01,
  cost_running_flat = 3.8,
  slope_equation = "original"
)
}
\arguments{
\item{data}{A data frame containing sprint alactic power estimations (default: \code{graubner_nixdorf_sprints}).}

\item{mu_al}{A double representing the peak of the log-normal distribution. Default is 1.75.}

\item{sigma_al}{A double representing the decay of the log-normal distribution. Default is 1.5.}

\item{line_color}{Color for model line (default: "purple").}

\item{point_color}{Color for data points (default: "darkorchid4").}

\item{mu}{A double. Parameter setting the peak of the log-normal distribution.Default is -0.4}

\item{sigma}{A double. Parameter setting the decay of the log-normal distribution. Default is 1}

\item{k1}{A double. Time constant of the first rising exponential (s). Default is 2.75}

\item{k2}{A double. Time constant of the second decaying exponential (s). Default is 35}

\item{cost_running_flat}{a numeric value representing the cost of running on a flat surface (default for this function is 3.8 J/kg/m, as used in Briand et al. 2025)}

\item{slope_equation}{a string indicating the slope equation to use, either "original" or "extended". The "original" equation is based on the equation relating running at constant speed on different slopes porposed by Minetti et al. (2002), while the "extended" equation is based on similar equation for extended slopes proposed by Minetti and Pavei (2018).}
}
\value{
A ggplot object with overlaid male and female model fits and alactic power over different running durations.
}
\description{
Provides observed alactic power values and model predictions
for male and female athletes using the log-normal decay model.The function using the graubner_nixdorf_sprints data set
generates the same Figure as the one presented in Briand et al. 2025.
}
\examples{
plot_sprint_alactic_duration_briand_article()
}
