% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sprint_energetics_exploration.R
\name{plot_sprint_approx_power_dist}
\alias{plot_sprint_approx_power_dist}
\title{Sprint Approximate Power Distribution Plot}
\usage{
plot_sprint_approx_power_dist(sprint_approx_power_distribution)
}
\arguments{
\item{sprint_approx_power_distribution}{A tibble with the following columns: time (s), power (W/kg), power_alactic (W/kg), power_lactic (W/kg), power_aerobic (W/kg)}
}
\value{
A ggplot object representing the approximate power distribution of the three metabolic pathways contribution to the total power during a sprint.
}
\description{
Plots the approximate power distribution of the three metabolic pathways contribution
to the total power during a sprint.
}
\examples{

# Extract the data for the 100 m
men_100 <- graubner_nixdorf_sprints |>
  dplyr::filter(event == "Men's 100 m")


# Get the sprint motion data for both men and women

sprint_data <- sprint_motion_model_data(
  mean_velocity_splits = men_100$velocity,
  time_splits = men_100$splits,
  distance = men_100$distance,
  reaction_time = men_100$reaction_time[1],
  maximal_velocity = men_100$maximal_velocity[1]
)

sprint_approx_power_distributions <- sprint_approx_power_distributions(sprint_data,
  maximal_aerobic_power = 24.5,
  basal_metabolic_rate = 1.2
)


plot_sprint_approx_power_dist(sprint_approx_power_distributions)

}
