% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sprint_bioenergetic_model.R
\name{plot_sprint_bioenergetic_model}
\alias{plot_sprint_bioenergetic_model}
\title{Plot Sprint Bioenergetic Model Power}
\usage{
plot_sprint_bioenergetic_model(
  sprint_motion_data,
  maximal_aerobic_power = 24.5,
  scale_max_power = 160
)
}
\arguments{
\item{sprint_motion_data}{A tibble with the following columns: time (s), velocity (m/s), acceleration (m/s^2), distance (m), cost of running (J/kg/m) and power (W/kg).}

\item{maximal_aerobic_power}{A double representing the maximal aerobic power (W/kg) of the athlete. Default is 24.5 W/kg.}

\item{scale_max_power}{A double representing the maximum y-axis value for the plot. Default is 160 W/kg.}
}
\value{
A ggplot object comparing the modeled metabolic power from the sprint bioenergetic model to the observed metabolic power from the sprint motion data.
}
\description{
Plot Sprint Bioenergetic Model Power
}
\examples{

#' # Extract Bolt's 100 m data from Graubner and Nixdorf data set.
bolt_100m <- graubner_nixdorf_sprints |>
  dplyr::filter(
    athlete == "Bolt",
    event == "Men's 100 m"
  )

# Compute sprint motion data

bolt_100m_motion_data <- sprint_motion_model_data(
  mean_velocity_splits = bolt_100m$velocity,
  time_splits = bolt_100m$splits,
  distance = bolt_100m$distance,
  reaction_time = bolt_100m$reaction_time[1],
  maximal_velocity = bolt_100m$maximal_velocity[1]
)


plot_sprint_bioenergetic_model(
sprint_motion_data = bolt_100m_motion_data
)



}
