% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sprint_lactic_duration_model.R
\name{plot_sprint_lactic_duration}
\alias{plot_sprint_lactic_duration}
\title{Plot Sprint Lactic Energy vs Duration with Model Overlay}
\usage{
plot_sprint_lactic_duration(
  lactic_energy_duration,
  t1 = 20,
  t2 = 1500,
  linetype = "solid",
  line_color = "#457B9D",
  point_color = "darkblue",
  point_shape = 16
)
}
\arguments{
\item{lactic_energy_duration}{A data frame or tibble containing columns \code{duration} (in seconds)
and \code{lactic_energy} (in J/kg), such output can be obtained through the use of Briand et al.' (2025) sprint bioenergetic model
and of the \code{sprint_lactic_energy_duration} function .}

\item{t1}{A numeric value representing the time constant for the first exponential decay (default is 20 seconds).}

\item{t2}{A numeric value representing the time constant for the second exponential decay (default is 1500 seconds).}

\item{linetype}{Line type used for the model curve (default = \code{"solid"}). Accepts any valid ggplot2 line type.}

\item{line_color}{Color of the model prediction line (default = \code{"#457B9D"}).}

\item{point_color}{Color of the observed data points (default = \code{"darkblue"}).}

\item{point_shape}{Shape of the observed data points (default = 16, i.e., solid circle).}
}
\value{
A \code{ggplot} object showing the observed lactic energy-duration data and the fitted model curve.
}
\description{
Creates a ggplot visualizing observed lactic energy expenditure as a function of sprint duration,
overlaid with a model prediction curve based on fitted lactic energy duration model.
}
\examples{
# Example with sample sprint data

data <- sprint_lactic_energy_duration_graubner_nixdorf()

plot_sprint_lactic_duration(data)

}
