% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sprint_lactic_duration_model.R
\name{plot_sprint_lactic_duration_briand_article}
\alias{plot_sprint_lactic_duration_briand_article}
\title{Plot lactic energy-duration model fits for male and female sprinters}
\usage{
plot_sprint_lactic_duration_briand_article(
  data = graubner_nixdorf_sprints,
  k1 = 20,
  k2 = 2000,
  line_color = "#457B9D",
  point_color = "darkblue"
)
}
\arguments{
\item{data}{A data frame containing sprint lactic energy estimations (default: \code{graubner_nixdorf_sprints}).}

\item{k1}{rising constant of the bi-exponential model (default = 20 s), describing lactic energy behavior over duration.}

\item{k2}{decaying constant of the bi-exponential model (default = 2000 s), describing lactic energy behavior over duration.}

\item{line_color}{Color for model line (default: "#457B9D").}

\item{point_color}{Color for data points (default: "darkblue").}
}
\value{
A ggplot object with overlaid male and female model fits and lactic energy over different running durations.
}
\description{
Provides observed lactic energy values and model predictions
for male and female athletes using the bi-exponential lactic energy-duration model.The function using the \code{graubner_nixdorf_sprints} data set
generates the same Figure as the one presented in Briand et al. 2025.
}
\examples{
plot_sprint_lactic_duration_briand_article()
}
