% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sprint_lactic_duration_model.R
\name{plot_sprint_lactic_duration_kindermann}
\alias{plot_sprint_lactic_duration_kindermann}
\title{Plot lactic energy-duration model fits on Kindermann's (1977) accumulated blood lactate data}
\usage{
plot_sprint_lactic_duration_kindermann(
  data = kindermann_lactate,
  t1 = 20,
  t2 = 1500,
  line_color = "#457B9D",
  point_color = "darkblue"
)
}
\arguments{
\item{data}{A data frame containing \code{accumulated_lactate} in mmol/L associated with different running \code{duration} in s (default: \code{kindermann_lactate}).}

\item{t1}{rising constant of the bi-exponential model (default = 20 s), describing lactic energy behavior over duration.}

\item{t2}{decaying constant of the bi-exponential model (default = 1500 s), describing lactic energy behavior over duration.}

\item{line_color}{Color for model line (default: "#457B9D").}

\item{point_color}{Color for data points (default: "darkblue").}
}
\value{
A ggplot object with overlaid male and female model fits and lactic energy over different running durations.
}
\description{
Provides observed lactic energy values and model predictions
for male and female athletes using the bi-exponential lactic energy-duration model. Lactic energy
is estimated using the conversion function \code{convert_acc_lactate_to_lactic_energy}. The accumulated blood lactate values are
provided in the \code{kindermann_lactate} data set. The function
generates a similar Figure as the one presented in Briand et al. 2025.
}
\examples{
plot_sprint_lactic_duration_kindermann()
}
