% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sprint_motion_analysis.R
\name{plot_sprint_motion_analysis}
\alias{plot_sprint_motion_analysis}
\title{Plot Sprint Motion Analysis}
\usage{
plot_sprint_motion_analysis(
  mean_velocity_splits,
  time_splits,
  distance,
  reaction_time,
  maximal_velocity = NA,
  dt = 0.01,
  cost_running_flat = 3.6,
  slope_equation = "original",
  color_distance = "darkgreen",
  color_velocity = "darkred",
  color_acceleration = "blue",
  color_power = "purple"
)
}
\arguments{
\item{mean_velocity_splits}{A vector with mean velocity splits over each distance interval (m/s)}

\item{time_splits}{A vector with time splits over each distance interval (s)}

\item{distance}{A vector with the distances at which time splits were measured (m)}

\item{reaction_time}{A double with the reaction time measured on the starting blocks (s)}

\item{maximal_velocity}{A double representing the maximal velocity (in m/s)}

\item{dt}{Time step of the model. Default is set at 0.01 s}

\item{cost_running_flat}{a numeric value representing the cost of running on a flat surface (default is 3.6 J/kg/m, as suggested in Minetti et al. 2002)}

\item{slope_equation}{a string indicating the slope equation to use, either "original" or "extended". The "original" equation is based on the equation relating running at constant speed on different slopes porposed by Minetti et al. (2002), while the "extended" equation is based on similar equation for extended slopes proposed by Minetti and Pavei (2018).}

\item{color_distance}{A string with the color of the distance plot's line and points. Default is set at "darkgreen"}

\item{color_velocity}{A string with the color of the velocity plot's line and points. Default is set at "darkred"}

\item{color_acceleration}{A string with the color of the acceleration plot's line and points. Default is set at "blue"}

\item{color_power}{A string with the color of the power plot's line and points. Default is set at "purple"}
}
\value{
A ggplot object
}
\description{
Plot of modeled sprint distance, velocity, acceleration and power as a function of time.
}
\examples{
plot_sprint_motion_analysis(
  mean_velocity_splits = c(0, 5.77, 9.99),
  time_splits = c(0, 1.88, 2.88),
  distance = c(0, 10, 20),
  reaction_time = 0.146,
  maximal_velocity = 12.34
)

}
