% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_motion_computations.R
\name{predicted_maximal_velocity}
\alias{predicted_maximal_velocity}
\title{Predicted Sprint Maximal Velocity}
\usage{
predicted_maximal_velocity(maximal_velocity, tau, time_maximal_velocity)
}
\arguments{
\item{maximal_velocity}{A double representing the maximal velocity (in m/s)}

\item{tau}{A double representing the time constant of the exponential rise in velocity (in s)}

\item{time_maximal_velocity}{A double representing the time at which the acceleration phase ends (in s)}
}
\value{
A double representing the predicted maximal velocity (in m/s)
}
\description{
Predicted Sprint Maximal Velocity
}
\examples{
predicted_maximal_velocity(10, 1.5, 5)
}
