% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runrgetics-package.R
\docType{package}
\name{runrgetics-package}
\alias{runrgetics}
\alias{runrgetics-package}
\title{runrgetics: Analyze and Quantify Running Energetics from Velocity Data}
\description{
A comprehensive toolkit for analyzing running energetics from velocity data. Functions are to calculate running power, estimate metabolic contributions of sprint running (aerobic, anaerobic alactic, and anaerobic lactic systems) and estimate anaerobic alactic and lactic capacities from a set of sprint running performances.analyze sprint performance. Designed for sports scientists,coaches, and researchers, it enables indirect quantification of running energetics using a method initially proposed by di Prampero (2005).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JrmyBriand/runrgetics}
  \item Report bugs at \url{https://github.com/JrmyBriand/runrgetics/issues}
}

}
\author{
\strong{Maintainer}: Jeremy Briand \email{jeremy.briand@umontreal.ca}

}
\keyword{internal}
