% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_style.R
\name{scale_colour_runrgetics}
\alias{scale_colour_runrgetics}
\title{runrgetics discrete colour scale}
\usage{
scale_colour_runrgetics(...)
}
\arguments{
\item{...}{Passed to \code{\link[ggplot2:scale_manual]{ggplot2::scale_colour_manual()}}.}
}
\value{
A ggplot2 colour scale.
}
\description{
Convenience \code{\link[ggplot2:scale_manual]{ggplot2::scale_colour_manual()}} wrapper using \code{\link[=runrgetics_pal]{runrgetics_pal()}}.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg, colour = factor(cyl))) +
  geom_point() +
  scale_colour_runrgetics() +
  theme_runrgetics()
}
