% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sprint_motion_analysis.R
\name{sex_comparison_plot_velocity}
\alias{sex_comparison_plot_velocity}
\title{Sex comparison velocity plot}
\usage{
sex_comparison_plot_velocity(
  sprint_data_male,
  sprint_data_female,
  observed_data_male,
  observed_data_female,
  color = "darkgreen"
)
}
\arguments{
\item{sprint_data_male}{A data frame with sprint motion model data extracted from males data}

\item{sprint_data_female}{A data frame with sprint motion model data extracted from females data}

\item{observed_data_male}{A data frame with columns "splits", "distance", "velocity" and "reaction_time" as in \code{\link{graubner_nixdorf_sprints}} data set}

\item{observed_data_female}{A data frame with columns "splits", "distance", "velocity" and "reaction_time" as in \code{\link{graubner_nixdorf_sprints}} data set}

\item{color}{A string with the color of the plot's line and points. Default is set at "darkgreen"}
}
\value{
A ggplot object
}
\description{
Generates plots to compare velocity as a function of time
for both male and females as presented in Briand et al. (2025), using
sprint data from Graubner and Nixdorf (2011).
}
\examples{

# Extract the data for the 100 m
men_100 <- graubner_nixdorf_sprints  |>
  dplyr::filter(event == "Men's 100 m")

women_100 <- graubner_nixdorf_sprints  |>
  dplyr::filter(event == "Women's 100 m")

# Get the sprint motion data for both men and women

sprint_data_men_100 <- sprint_motion_model_data(
  mean_velocity_splits = men_100$velocity,
  time_splits = men_100$splits,
  distance = men_100$distance,
  reaction_time = men_100$reaction_time[1],
  maximal_velocity = men_100$maximal_velocity[1]
)

sprint_data_women_100 <- sprint_motion_model_data(
  mean_velocity_splits = women_100$velocity,
  time_splits = women_100$splits,
  distance = women_100$distance,
  reaction_time = women_100$reaction_time[1],
  maximal_velocity = women_100$maximal_velocity[1]
)


#Generate the plots

sex_comparison_plot_velocity(
  sprint_data_male = sprint_data_men_100,
  sprint_data_female = sprint_data_women_100,
  observed_data_male = men_100,
  observed_data_female = women_100,
  color = "darkgreen") + ggplot2::ggtitle("Sex Comparison of the 100 m Instantaneous Velocity")
}
