% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_motion_computations.R
\name{sprint_acc_distance_model}
\alias{sprint_acc_distance_model}
\title{Instantaneous Distance During Running Acceleration}
\usage{
sprint_acc_distance_model(time, tau, maximal_velocity)
}
\arguments{
\item{time}{A vector of time points (in s)}

\item{tau}{A double representing the time constant of the exponential rise in velocity (in s)}

\item{maximal_velocity}{A double representing the maximal velocity (in m/s)}
}
\value{
A vector of instantaneous distances (in m)
}
\description{
Estimates the instantaneous distance during the acceleration phase of a sprint run. The model is based on an exponential rise in velocity.
}
\examples{
sprint_acc_distance_model(1:10, 0.5, 10)
}
