% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_energetics_exploration.R
\name{sprint_acceleration_approx_lactic_power}
\alias{sprint_acceleration_approx_lactic_power}
\title{Sprint Approximate Anaerobic Lactic Power (Acceleration)}
\usage{
sprint_acceleration_approx_lactic_power(time, maximal_lactic_power, k_an = 2)
}
\arguments{
\item{time}{A vector of time points (s)}

\item{maximal_lactic_power}{A double corresponding to the maximal anaerobic lactic power (W/kg).}

\item{k_an}{A double corresponding to th time constant (s) of the exponential rise to the approximate maximal lactic power. Default is 2 s.}
}
\value{
A vector of the approximate anaerobic lactic power (W/kg) at each time point
}
\description{
Sprint Approximate Anaerobic Lactic Power (Acceleration)
}
\examples{

sprint_acceleration_approx_lactic_power(time = seq(0, 10, by = 0.1), maximal_lactic_power = 10)

}
