% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_motion_computations.R
\name{sprint_acceleration_model}
\alias{sprint_acceleration_model}
\title{Instantaneous Acceleration During Sprint Running (Acceleration and Deceleration phases)}
\usage{
sprint_acceleration_model(
  time,
  time_maximal_velocity,
  maximal_velocity,
  tau,
  decel_rate
)
}
\arguments{
\item{time}{A vector of time points (in s)}

\item{time_maximal_velocity}{A double, time at which maximal velocity is reached (in s)}

\item{maximal_velocity}{A double, maximal velocity attained during the sprint (in m/s)}

\item{tau}{A double, time constant of the exponential rise in velocity (in s)}

\item{decel_rate}{A double, rate of deceleration (in m/s^2)}
}
\value{
A vector of instantaneous accelerations (in m/s^2)
}
\description{
Estimates the instantaneous acceleration during the acceleration and deceleration phases of a sprint run.
The calculation is based on an exponential rise in velocity during the acceleration phase and a constant rate of deceleration during the deceleration phase.
}
\examples{
sprint_acceleration_model(1:10, 5, 10, 1.5, 0.1)

}
