% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_alactic_energy_duration.R
\name{sprint_alactic_duration_model}
\alias{sprint_alactic_duration_model}
\title{Sprint Alactic Duration Model}
\usage{
sprint_alactic_duration_model(
  duration,
  alactic_capacity,
  mu_al = 1.75,
  sigma_al = 1.5
)
}
\arguments{
\item{duration}{A numeric vector representing the duration of the sprint (in seconds).}

\item{alactic_capacity}{A numeric value representing the alactic capacity (in J/kg).}

\item{mu_al}{A double representing the peak of the log-normal distribution. Default is 1.75.}

\item{sigma_al}{A double representing the decay of the log-normal distribution. Default is 1.5.}
}
\value{
A numeric vector representing the average alactic power (in W/kg) as a function of running duration.
}
\description{
Computes the alactic power as a function of duration using a log-normal distribution model.
Other than duration, the alactic capacity is the only parameter of the model. The alactic capacity corresponds to the maximal energy available through the alactic energy pathway.
}
\examples{

alactic_capacity <- 350 # Example alactic capacity in J/kg
duration <- seq(0.1, 10, by = 0.1) # Example durations from 0.1 to 10 seconds

alactic_power <- sprint_alactic_duration_model(duration, alactic_capacity)
alactic_power


}
