% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_alactic_energy_duration.R
\name{sprint_alactic_energy_duration_graubner_nixdorf}
\alias{sprint_alactic_energy_duration_graubner_nixdorf}
\title{Sprint Alactic Energy and Power Over Running Duration (Graubner and Nixdorf dataset)}
\usage{
sprint_alactic_energy_duration_graubner_nixdorf(
  data = graubner_nixdorf_sprints,
  athlete_sex = "male",
  mu = -0.4,
  sigma = 1,
  k1 = 2.75,
  k2 = 35,
  dt = 0.01,
  cost_running_flat = 3.8,
  slope_equation = "original"
)
}
\arguments{
\item{data}{A tibble with the following: distance (m), splits (s), velocity (m/s), reaction_time (s), maximal_velocity (m/s) and event (character). Default is Graubner and Nixdorf (2009) sprint data.}

\item{athlete_sex}{A character corresponding the athlete sex for the analysis. Default is "male" and computes energy and power over the Graubner and Nixdorf's (2011) men's 100, 200 and 400 m performances. It also sets the maximal aerobic power to 24.5 W/kg. Similarly, setting athlete_sex to "female" performs the computation on women's performances and sets maximal aerobic power to 21 W/kg.}

\item{mu}{A double. Parameter setting the peak of the log-normal distribution.Default is -0.4}

\item{sigma}{A double. Parameter setting the decay of the log-normal distribution. Default is 1}

\item{k1}{A double. Time constant of the first rising exponential (s). Default is 2.75}

\item{k2}{A double. Time constant of the second decaying exponential (s). Default is 35}

\item{cost_running_flat}{a numeric value representing the cost of running on a flat surface (default for this function is 3.8 J/kg/m, as used in Briand et al. 2025)}

\item{slope_equation}{a string indicating the slope equation to use, either "original" or "extended". The "original" equation is based on the equation relating running at constant speed on different slopes porposed by Minetti et al. (2002), while the "extended" equation is based on similar equation for extended slopes proposed by Minetti and Pavei (2018).}
}
\value{
A tibble with duration (s), alactic_energy (J/kg), alactic_power (W/kg).
}
\description{
Computes the alactic energy and average power over the duration of a sprint for the Graubner and Nixdorf dataset.
The default data is the \code{graubner_nixdorf_sprints} dataset, which contains sprint performances of elite athletes.
The \code{athlete_sex} parameter enables to compute alactic energy and power over running duration for either male or female 100, 200 and 400 m events.
}
\examples{

sprint_alactic_energy_duration_graubner_nixdorf()

}
