% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_energetics_exploration.R
\name{sprint_approx_alactic_power_model}
\alias{sprint_approx_alactic_power_model}
\title{Sprint Alactic Power Model (on Approximate Alactic Power Distribution)}
\usage{
sprint_approx_alactic_power_model(time, maximal_alactic_power, mu, sigma)
}
\arguments{
\item{time}{A vector of time points (s)}

\item{maximal_alactic_power}{A double. Maximal alactic power over the race duration (W/kg).}

\item{mu}{A double. Parameter setting the peak of the log-normal distribution.}

\item{sigma}{A double. Parameter setting the decay of the log-normal distribution.}
}
\value{
A vector of the approximate alactic power (W/kg) at each time point
}
\description{
Estimates the alactic power at each time point using the alactic power model based on a log-normal function and
the approximate alactic power estimations through the course of the sprint.
}
\examples{

maximal_alactic_power <- 160
mu <- 1
sigma <- 1

time <- seq(1,10, 0.1)

sprint_approx_alactic_power_model(time, maximal_alactic_power, mu, sigma)


}
