% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_energetics_exploration.R
\name{sprint_approx_lactic_power_model}
\alias{sprint_approx_lactic_power_model}
\title{Sprint Lactic Power Model (on Approximate Lactic Power Distribution)}
\usage{
sprint_approx_lactic_power_model(time, maximal_lactic_power, k1, k2)
}
\arguments{
\item{time}{A vector of time points (s)}

\item{maximal_lactic_power}{A double. Maximal lactic power over the race duration (W/kg).}

\item{k1}{A double. Time constant of the first rising exponential (s).}

\item{k2}{A double. Time constant of the second decaying exponential (s).}
}
\value{
A vector of the approximate lactic power (W/kg) at each time point
}
\description{
Estimates lactic power at each time point using the lactic power model based on a bi_exponential
and approximate lactic power through the course of the sprint
}
\examples{

maximal_lactic_power <- 60
k1 <- 2.5
k2 <- 35

time <- seq(1, 10, 0.1)

sprint_approx_lactic_power_model(time, maximal_lactic_power, k1, k2)

}
