% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_energetics_exploration.R
\name{sprint_approx_power_distributions}
\alias{sprint_approx_power_distributions}
\title{Sprint Approximate Power Distributions}
\usage{
sprint_approx_power_distributions(
  sprint_motion_data,
  maximal_aerobic_power = 24.5,
  basal_metabolic_rate = 1.2
)
}
\arguments{
\item{sprint_motion_data}{A tibble with the following columns: time, velocity, acceleration, distance, cost of running and power}

\item{maximal_aerobic_power}{A double. Maximal aerobic power or MAP (W/kg). Default is 24.5 W/kg corresponding to World class male sprinter observed MAP. The equivalent for female is 21 W/kg}

\item{basal_metabolic_rate}{A double. Basal metabolic rate (W/kg). Default is 1.2 W/kg}
}
\value{
A tibble with the following columns: time, velocity, acceleration, distance, cost of running, power, power_aerobic, power_anaerobic, power_lactic, power_alactic
}
\description{
Computes approximate aerobic, lactic and alactic power distributions from sprint motion data.
}
\examples{

# Extract the data for the 100 m
men_100 <- graubner_nixdorf_sprints |>
  dplyr::filter(event == "Men's 100 m")


# Get the sprint motion data for both men and women

sprint_data <- sprint_motion_model_data(
  mean_velocity_splits = men_100$velocity,
  time_splits = men_100$splits,
  distance = men_100$distance,
  reaction_time = men_100$reaction_time[1],
  maximal_velocity = men_100$maximal_velocity[1]
)

sprint_approx_power_distributions(sprint_data,
  maximal_aerobic_power = 24.5,
  basal_metabolic_rate = 1.2
)

}
