% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_bioenergetic_model.R
\name{sprint_bioenergetic_model}
\alias{sprint_bioenergetic_model}
\title{Sprint Bioenergetic Model}
\usage{
sprint_bioenergetic_model(
  time,
  maximal_alactic_power,
  maximal_lactic_power,
  mu = -0.4,
  sigma = 1,
  k1 = 2.75,
  k2 = 35,
  maximal_aerobic_power = 24.5,
  output = "total power"
)
}
\arguments{
\item{time}{A numeric vector representing time points (in seconds) at which to calculate the power output.}

\item{maximal_alactic_power}{A double corresponding to the maximal alactic power attained over the sprint (W/kg)}

\item{maximal_lactic_power}{A double corresponding to the maximal lactic power attained over the sprint (W/kg)}

\item{mu}{A double. Parameter setting the peak of the log-normal distribution.Default is -0.4}

\item{sigma}{A double. Parameter setting the decay of the log-normal distribution. Default is 1}

\item{k1}{A double. Time constant of the first rising exponential (s). Default is 2.75}

\item{k2}{A double. Time constant of the second decaying exponential (s). Default is 35}

\item{maximal_aerobic_power}{A double. Maximal aerobic power (W/kg). Default is 24.5}

\item{output}{A string. The type of power output to return. Options are "total power", "alactic power", "lactic power", and "aerobic power". Default is "total power".}
}
\value{
A double representing the power output (W/kg) at a given time point.
}
\description{
Sprint Bioenergetic Model
}
\examples{


time <- seq(0, 30, by = 0.01)
maximal_alactic_power <- 100
maximal_lactic_power <- 50


sprint_bioenergetic_model(time, maximal_alactic_power, maximal_lactic_power)

}
