% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_bioenergetic_model_comparison_tables.R
\name{sprint_bioenergetic_model_gof_metrics}
\alias{sprint_bioenergetic_model_gof_metrics}
\title{Sprint Bioenergetic Model Goodness of Fit Metrics}
\usage{
sprint_bioenergetic_model_gof_metrics(
  sprint_power_data,
  event,
  mu = -0.4,
  sigma = 1,
  k1 = 2.75,
  k2 = 35,
  maximal_aerobic_power = 24.5,
  dt = 0.01
)
}
\arguments{
\item{sprint_power_data}{A tibble with at least the following columns: time (s), power (W/kg), power_mod (W/kg), corresponding respectively to the power estimated by di Prampero et al.'s (2005, 2018) equivalent slope approach, and the power estimated using Briand et al.'s (2025) sprint bioenergetic model.}

\item{event}{A character corresponding to the name of the events from which the sprint power data was derived.}

\item{mu}{A double. Parameter setting the peak of the log-normal distribution.Default is -0.4}

\item{sigma}{A double. Parameter setting the decay of the log-normal distribution. Default is 1}

\item{k1}{A double. Time constant of the first rising exponential (s). Default is 2.75}

\item{k2}{A double. Time constant of the second decaying exponential (s). Default is 35}

\item{maximal_aerobic_power}{A double corresponding to the estimated maximal aerobic power of the sprinter (W/kg). Default is 24.5.}

\item{dt}{Time step of the model. Default is set at 0.01 s}
}
\value{
A tibble with the following columns: event, adjusted_Rsquared, energy_total, energy_percent_diff, estimated_max_power, percentage_difference_max_power, distance_modeled_recovered, distance_percent_diff.
}
\description{
Computes a set of metrics that can be used to assess the sprint bioenergetic model's goodness of fit to the observed metabolic power data.
These metrics include adjusted R-squared, total energy contribution, percentage difference in energy contribution, estimated maximum metabolic power,
percentage difference in maximum metabolic power, modeled recovered distance, and percentage difference in recovered distance. The metrics were reported in
Briand et al. (2025) and are used to compare the sprint bioenergetic model to the observed metabolic power data from di Prampero et al. (2005, 2018) derived from Graubner and Nixdorf (2011) data.
}
\examples{

# Extract Bolt's 100 m data from Graubner and Nixdorf data set.
bolt_100m <- graubner_nixdorf_sprints |>
  dplyr::filter(
    athlete == "Bolt",
    event == "Men's 100 m"
  )

# Compute sprint motion data

bolt_100m_motion_data <- sprint_motion_model_data(
  mean_velocity_splits = bolt_100m$velocity,
  time_splits = bolt_100m$splits,
  distance = bolt_100m$distance,
  reaction_time = bolt_100m$reaction_time[1],
  maximal_velocity = bolt_100m$maximal_velocity[1]
)


# compute the power estimated from the sprint bioenergetic model

bolt_modeled_data <- sprint_bioenergetic_model_data(bolt_100m_motion_data)

sprint_bioenergetic_model_gof_metrics(bolt_modeled_data, event = "Men's 100 m")

}
