% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_bioenergetic_model_fit_diagnostic.R
\name{sprint_bioenergetic_model_R2}
\alias{sprint_bioenergetic_model_R2}
\title{Sprint Bioenergetic Model R-squared}
\usage{
sprint_bioenergetic_model_R2(
  sprint_motion_data,
  mu = -0.4,
  sigma = 1,
  k1 = 2.75,
  k2 = 35,
  maximal_aerobic_power = 24.5
)
}
\arguments{
\item{sprint_motion_data}{A tibble with the following columns: time (s), velocity (m/s), acceleration (m/s^2), distance (m), cost of running (J/kg/m) and power (W/kg).}

\item{mu}{A double. Parameter setting the peak of the log-normal distribution.Default is -0.4}

\item{sigma}{A double. Parameter setting the decay of the log-normal distribution. Default is 1}

\item{k1}{A double. Time constant of the first rising exponential (s). Default is 2.75}

\item{k2}{A double. Time constant of the second decaying exponential (s). Default is 35}

\item{maximal_aerobic_power}{A double corresponding to the estimated maximal aerobic power of the sprinter (W/kg). Default is 24.5.}
}
\value{
A double with the R-squared of the model fit.
}
\description{
Computes the R-squared of the sprint bioenergetic model non-linear least squares fit. The fit is based on sprint metabolic power data derived from
available time splits, using the approach proposed by di Prampero et al. (2005, 2018) and as described in Briand et al. (2025).
}
\examples{

# Extract Bolt's 100 m data from Graubner and Nixdorf data set.
bolt_100m <- graubner_nixdorf_sprints |>
  dplyr::filter(
    athlete == "Bolt",
    event == "Men's 100 m"
  )

# Compute sprint motion data

bolt_100m_motion_data <- sprint_motion_model_data(
  mean_velocity_splits = bolt_100m$velocity,
  time_splits = bolt_100m$splits,
  distance = bolt_100m$distance,
  reaction_time = bolt_100m$reaction_time[1],
  maximal_velocity = bolt_100m$maximal_velocity[1]
)


sprint_bioenergetic_model_R2(bolt_100m_motion_data)

}
