% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_bioenergetic_model_comparison_tables.R
\name{sprint_briand_article_gof_table}
\alias{sprint_briand_article_gof_table}
\title{Sprint Bioenergetic Model Goodness of Fit Table}
\usage{
sprint_briand_article_gof_table(
  data = graubner_nixdorf_sprints,
  mu = -0.4,
  sigma = 1,
  k1 = 2.75,
  k2 = 35,
  dt = 0.01
)
}
\arguments{
\item{data}{A tibble with the following: distance (m), splits (s), velocity (m/s), reaction_time (s), maximal_velocity (m/s) and event (character). Default is Graubner and Nixdorf (2009) sprint data.}

\item{mu}{A double. Parameter setting the peak of the log-normal distribution.Default is -0.4}

\item{sigma}{A double. Parameter setting the decay of the log-normal distribution. Default is 1}

\item{k1}{A double. Time constant of the first rising exponential (s). Default is 2.75}

\item{k2}{A double. Time constant of the second decaying exponential (s). Default is 35}

\item{dt}{A double representing the time step at which power is provided (in seconds). Default is 0.01 seconds.}
}
\value{
A tinytable object with the following columns: event, adjusted_Rsquared, energy_total, energy_percent_diff, estimated_max_power, percentage_difference_max_power, distance_modeled_recovered, distance_percent_diff.
}
\description{
Computes a table of metrics assessing the sprint bioenergetic model's goodness of fit on metabolic power data computed
using di Prampero et al.'s (2005, 2018) equivalent slope approach, as presented
in Briand et al. (2025). The table presents the metrics for the men's and women's 100, 200 and 400 m
provided in Graubner and Nixdorf's (2011) dataset. The table includes metrics such as adjusted R-squared, total energy contribution, percentage difference in energy contribution, estimated maximum metabolic power, percentage difference in maximum metabolic power, modeled recovered distance, and percentage difference in recovered distance. The metrics are computed for each event in the provided data set.
}
\examples{

sprint_briand_article_gof_table()

}
