% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_motion_computations.R
\name{sprint_dec_distance_model}
\alias{sprint_dec_distance_model}
\title{Instantaneous Distance During Sprint Running Deceleration Phase}
\usage{
sprint_dec_distance_model(
  time,
  maximal_velocity,
  time_maximal_velocity,
  distance_maximal_velocity,
  decel_rate
)
}
\arguments{
\item{time}{A vector of time points (in s)}

\item{maximal_velocity}{A double representing the maximal velocity (in m/s)}

\item{time_maximal_velocity}{A double representing the time at which maximal velocity is reached (in s)}

\item{distance_maximal_velocity}{A double representing the distance at which maximal velocity is reached (in m)}

\item{decel_rate}{A double representing the rate of deceleration (in m/s^2)}
}
\value{
A vector of instantaneous distances (in m)
}
\description{
Estimates the instantaneous distance during the deceleration phase of a sprint run. The model is based on a constant rate of deceleration.
}
\examples{
sprint_dec_distance_model(6:10, 10, 6, 60, 0.1)
}
