% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_bioenergetic_model_comparison_tables.R
\name{sprint_energy_cont_briand_article_table}
\alias{sprint_energy_cont_briand_article_table}
\title{Sprint Metabolic Pathways Energy Contributions Table}
\usage{
sprint_energy_cont_briand_article_table(
  data = graubner_nixdorf_sprints,
  mu = -0.4,
  sigma = 1,
  k1 = 2.75,
  k2 = 35,
  dt = 0.01,
  cost_running_flat = 3.8,
  slope_equation = "original"
)
}
\arguments{
\item{data}{A tibble with the following: distance (m), splits (s), velocity (m/s), reaction_time (s), maximal_velocity (m/s) and event (character). Default is Graubner and Nixdorf (2009) sprint data.}

\item{mu}{A double. Parameter setting the peak of the log-normal distribution.Default is -0.4}

\item{sigma}{A double. Parameter setting the decay of the log-normal distribution. Default is 1}

\item{k1}{A double. Time constant of the first rising exponential (s). Default is 2.75}

\item{k2}{A double. Time constant of the second decaying exponential (s). Default is 35}

\item{dt}{Time step of the model. Default is set at 0.01 s}

\item{cost_running_flat}{a numeric value representing the cost of running on a flat surface (default for this function is 3.8 J/kg/m, as used in Briand et al. 2025)}

\item{slope_equation}{a string indicating the slope equation to use, either "original" or "extended". The "original" equation is based on the equation relating running at constant speed on different slopes porposed by Minetti et al. (2002), while the "extended" equation is based on similar equation for extended slopes proposed by Minetti and Pavei (2018).}
}
\value{
A tinytable object with the following columns: event, alactic_energy, lactic_energy, aerobic_energy, total_energy, alactic_percentage, lactic_percentage, aerobic_percentage.
}
\description{
Computes a table of energy contributions of the alactic, lactic and aerobic pathways during sprint events using the sprint bioenergetic model.
The function provides the total energy expended during the sprint and the percentage contributions of each pathway to the total energy expenditure.
The table computes the energy contribution in the men's and women's 100, 200 and 400 m and corresponds to Table 3 from Briand et al. (2025).
}
\examples{

sprint_energy_cont_briand_article_table()

}
