% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_bioenergetic_model_comparison_tables.R
\name{sprint_energy_contributions}
\alias{sprint_energy_contributions}
\title{Sprint Event Metabolic Pathways Energy Contributions}
\usage{
sprint_energy_contributions(sprint_power_data, event_name)
}
\arguments{
\item{sprint_power_data}{A tibble with at least the following columns: time (s), power_alactic (W/kg), power_lactic (W/kg) and power_aerobic (W/kg), as well as power (W/kg), power_mod (W/kg), corresponding respectively to the power estimated by di Prampero et al.'s (2005, 2018) equivalent slope approach, and the power estimated using Briand et al.'s (2025) sprint bioenergetic model.}

\item{event_name}{A character corresponding to the name of the event from which the sprint power data was derived.}
}
\value{
A tibble with the following columns: event, alactic_energy, lactic_energy, aerobic_energy, total_energy, alactic_percentage, lactic_percentage, aerobic_percentage.
}
\description{
Computes the energy contributions of the alactic, lactic and aerobic pathways during a sprint event using the sprint bioenergetic model.
The function also provides the total energy expended during the sprint and the percentage contributions of each pathway to the total energy expenditure.
}
\examples{
# Extract Bolt's 100 m data from Graubner and Nixdorf data set.
bolt_100m <- graubner_nixdorf_sprints |>
  dplyr::filter(
    athlete == "Bolt",
    event == "Men's 100 m"
  )

# Compute sprint motion data

bolt_100m_motion_data <- sprint_motion_model_data(
  mean_velocity_splits = bolt_100m$velocity,
  time_splits = bolt_100m$splits,
  distance = bolt_100m$distance,
  reaction_time = bolt_100m$reaction_time[1],
  maximal_velocity = bolt_100m$maximal_velocity[1]
)


# compute the power estimated from the sprint bioenergetic model

bolt_modeled_data <- sprint_bioenergetic_model_data(bolt_100m_motion_data)

sprint_energy_contributions(bolt_modeled_data, event = "Men's 100 m")


}
