% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_extract_capacities_from_athlete_perf.R
\name{sprint_get_capacities_from_perf}
\alias{sprint_get_capacities_from_perf}
\title{Calculate Optimal Alactic and Lactic Capacities from Sprint Times}
\usage{
sprint_get_capacities_from_perf(
  performance_data,
  maximal_aerobic_power = 24.5,
  dt = 0.01,
  init_alactic = 350,
  init_lactic = 1000,
  lower_bounds = c(200, 1000),
  upper_bounds = c(400, 2000),
  mu = -0.4,
  sigma = 1,
  k1 = 2.75,
  k2 = 35,
  cost_running_flat = 3.6,
  slope_equation = "original"
)
}
\arguments{
\item{performance_data}{A tibble containing at least two columns:
\itemize{
\item distance: The race distance in meters
\item performance: The race time in seconds
}}

\item{maximal_aerobic_power}{A double. Maximal aerobic power (W/kg). Default is 24.5}

\item{dt}{Time step for numerical integration (default = 0.01)}

\item{init_alactic}{Initial guess for alactic capacity (default = 350)}

\item{init_lactic}{Initial guess for lactic capacity (default = 1000)}

\item{lower_bounds}{Numeric vector of length 2 specifying lower bounds for
alactic and lactic capacities respectively (default = c(200, 1000))}

\item{upper_bounds}{Numeric vector of length 2 specifying upper bounds for
alactic and lactic capacities respectively (default = c(400, 2000))}

\item{mu}{A double. Parameter setting the peak of the log-normal distribution.Default is -0.4}

\item{sigma}{A double. Parameter setting the decay of the log-normal distribution. Default is 1}

\item{k1}{A double. Time constant of the first rising exponential (s). Default is 2.75}

\item{k2}{A double. Time constant of the second decaying exponential (s). Default is 35}

\item{cost_running_flat}{a numeric value representing the cost of running on a flat surface (default is 3.6 J/kg/m, as suggested in Minetti et al. 2002)}

\item{slope_equation}{a string indicating the slope equation to use, either "original" or "extended". The "original" equation is based on the equation relating running at constant speed on different slopes porposed by Minetti et al. (2002), while the "extended" equation is based on similar equation for extended slopes proposed by Minetti and Pavei (2018).}
}
\value{
A named list containing:
\itemize{
\item alactic_capacity: Optimal alactic capacity
\item lactic_capacity: Optimal lactic capacity
\item convergence: Optimization convergence status
\item error: Final error value
}
}
\description{
Computes the optimal alactic and lactic capacities based on
sprint performances across different distances. It uses an optimization approach
to find the capacities that best explain the observed performances.
}
\examples{
performance_data <- tibble::tibble(
  distance = c(100, 200, 400),
  performance = c(9.43, 19.1, 43.9)
)
#sprint_get_capacities_from_perf(performance_data)

}
