% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_lactic_energy_duration.R
\name{sprint_lactic_duration_model}
\alias{sprint_lactic_duration_model}
\title{Sprint Lactic Energy Duration Model}
\usage{
sprint_lactic_duration_model(duration, lactic_capacity, k1 = 20, k2 = 2000)
}
\arguments{
\item{duration}{A numeric vector representing the duration of the sprint (in seconds).}

\item{lactic_capacity}{A numeric value representing the lactic capacity (in J/kg).}

\item{k1}{A double corresponding to the rising time-constant of the bi-exponential model. Default is 20.}

\item{k2}{A double representing to decaying time-constant of the bi-exponential model. Default is 2000.}
}
\value{
A numeric vector representing the average lactic energy (in J/kg) as a function of running duration.
}
\description{
Computes the lactic energy as a function of duration using a bi-exponential model.
Other than duration, the lactic capacity is the only parameter of the model. The lactic capacity corresponds to the maximal energy available through the lactic energy pathway.
}
\examples{

lactic_capacity <- 1350 # Example alactic capacity in J/kg
duration <- seq(0.1, 10, by = 0.1) # Example durations from 0.1 to 10 seconds

lactic_energy <- sprint_lactic_duration_model(duration, lactic_capacity)
lactic_energy


}
