% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_lactic_energy_duration.R
\name{sprint_lactic_duration_model_fit}
\alias{sprint_lactic_duration_model_fit}
\title{Sprint Lactic Duration Model Fit}
\usage{
sprint_lactic_duration_model_fit(lactic_energy_duration, k1 = 20, k2 = 2000)
}
\arguments{
\item{lactic_energy_duration}{A tibble with the following columns: duration (s), lactic_energy (J/kg), corresponding to the average lactic energy expenditure over running duration. This data is used to fit the lactic energy model.}

\item{k1}{A double corresponding to the rising time-constant of the bi-exponential model. Default is 20.}

\item{k2}{A double representing to decaying time-constant of the bi-exponential model. Default is 2000.}
}
\value{
A fitted model object of class "nls" representing the lactic energy model fitted to the lactic energy duration data.
}
\description{
Performs a non-linear least squares fitting of the lactic energy
model to the lactic energy over running duration data.
}
\examples{
men_lactic_energy <- sprint_lactic_energy_duration_graubner_nixdorf()
men_lactic_energy

 # Fit lactic model to energy

 sprint_lactic_duration_model_fit(men_lactic_energy)

}
