% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_lactic_energy_duration.R
\name{sprint_lactic_energy_duration}
\alias{sprint_lactic_energy_duration}
\title{Sprint Lactic Energy and Power Over Running Duration}
\usage{
sprint_lactic_energy_duration(sprint_power_data)
}
\arguments{
\item{sprint_power_data}{A tibble with a time column (time, in s) and instantaneous lactic power (power_lactic, in W/kg) over the sprint, as computed using the sprint bioenergetic model proposed by Briand et al. (2025).}
}
\value{
A tibble with the following columns: duration (s), lactic_energy (J/kg), lactic_power (W/kg).
}
\description{
Computes the lactic energy and average power over the duration of a sprint. The function is based on a provided sprint power data frame
containing instantaneous power contribution of the alactic, lactic and aerobic energy pathways. Such data frame can be derived using Briand et al.'s (2025)
sprint bioenergetic model
}
\examples{
# Extract Bolt's 100 m data from Graubner and Nixdorf data set.
bolt_100m <- graubner_nixdorf_sprints |>
  dplyr::filter(
    athlete == "Bolt",
    event == "Men's 100 m"
  )

# Compute sprint motion data

bolt_100m_motion_data <- sprint_motion_model_data(
  mean_velocity_splits = bolt_100m$velocity,
  time_splits = bolt_100m$splits,
  distance = bolt_100m$distance,
  reaction_time = bolt_100m$reaction_time[1],
  maximal_velocity = bolt_100m$maximal_velocity[1]
)


# compute the power estimated from the sprint bioenergetic model

bolt_modeled_data <- sprint_bioenergetic_model_data(bolt_100m_motion_data)


# compute lactic energy vs duration for Bolt's 100 m performance

 sprint_lactic_energy_duration(bolt_modeled_data)

}
