% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_lactic_model_formalisation.R
\name{sprint_lactic_model_gof_comp_table}
\alias{sprint_lactic_model_gof_comp_table}
\title{Formatted Table of Comparison of the Optimal and Formalized/Simplified Lactic Models Parameters and Goodness of Fit Metrics (On Berlin 2009 Sprint Performances)}
\usage{
sprint_lactic_model_gof_comp_table(
  data = graubner_nixdorf_sprints,
  dt = 0.01,
  cost_running_flat = 3.8,
  slope_equation = "original"
)
}
\arguments{
\item{data}{A tibble with the following: distance (m), splits (s), velocity (m/s), reaction_time (s), maximal_velocity (m/s) and event (character). Default is Graubner and Nixdorf (2009) sprint data.}

\item{dt}{Time step of the model. Default is set at 0.01 s}

\item{cost_running_flat}{a numeric value representing the cost of running on a flat surface (default for this function is 3.8 J/kg/m, as used in Briand et al. 2025)}

\item{slope_equation}{a string indicating the slope equation to use, either "original" or "extended". The "original" equation is based on the equation relating running at constant speed on different slopes porposed by Minetti et al. (2002), while the "extended" equation is based on similar equation for extended slopes proposed by Minetti and Pavei (2018).}
}
\value{
A tinytable comparing the optimal and formalized/simplified lactic models fitted parameters and goodness of fit metrics over different events: event (character), p_la_max (W/kg), k1 (s), k2 (s), AIC, BIC and R-squared.
}
\description{
Provides a formatted table comparing the optimal and formalized/simplified lactic models fitted parameters and goodness of fit metrics over different events. The optimal model
is a model where the maximal lactic power as well as k1 and k2 parameters of the bi-exponential power distribution are fitted to the data.
The formalized model is a model where the k1 and k2 parameters of the bi-exponential distribution are set at 2.75 and 35 s, respectively.
}
\examples{
sprint_lactic_model_gof_comp_table()

}
