% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sprint_motion_metrics.R
\name{sprint_maximum_metabolic_power}
\alias{sprint_maximum_metabolic_power}
\title{Maximum Metabolic Power}
\usage{
sprint_maximum_metabolic_power(sprint_motion_data, type = "power")
}
\arguments{
\item{sprint_motion_data}{A tibble with the following columns: time, velocity, acceleration, distance, cost of running and power. The maximum of the estimated power can also be computed by entering an extra column power_mod corresponding to the power derived from the sprint bioenergetic model.}

\item{type}{A string indicating the type of power to integrate. Default is "power" and refers to estimated metabolic power derived from di Prampero et al. equivalent slope approach. "power model" refers to the power estimated using the sprint bioenergetic model.}
}
\value{
A double with the maximal metabolic power (W/kg)
}
\description{
Computes the maximal metabolic power generated during a sprint event.The maximum of the estimated power using the sprint bioenergetic model
can also be computed by entering an extra column (power_mod) corresponding to the power derived from the sprint bioenergetic model.The \code{type} parameter has to be adjusted accordingly.
}
\examples{
# Extract the data for the 100 m
men_100 <- graubner_nixdorf_sprints |>
  dplyr::filter(event == "Men's 100 m")


# Get the sprint motion data for both men and women

sprint_data <- sprint_motion_model_data(
  mean_velocity_splits = men_100$velocity,
  time_splits = men_100$splits,
  distance = men_100$distance,
  reaction_time = men_100$reaction_time[1],
  maximal_velocity = men_100$maximal_velocity[1]
)

sprint_maximum_metabolic_power(sprint_data)

}
