% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_recover_motion_data.R
\name{sprint_modeled_distance_percentage_error}
\alias{sprint_modeled_distance_percentage_error}
\title{Sprint Recovered Distance Percentage Error}
\usage{
sprint_modeled_distance_percentage_error(sprint_power_data, dt = 0.01)
}
\arguments{
\item{sprint_power_data}{A tibble with at leats following columns: time (s), power (W/kg), corresponding to the power estimated by di Prampero et al.'s (2005, 2018) equivalent slope approach, power_mod (W/kg), corresponding to power estimated using Briand et al.'s (2025) sprint bioenergetic model.}

\item{dt}{a numeric value representing the time step at which power is provided (in seconds). Default is 0.01 seconds.}
}
\value{
A numeric value representing the percentage difference in distance recovered between the two methods.
}
\description{
Compares the distance recovered from the sprint power data using di Prampero et al.'s (2005, 2018) equivalent slope approach and Briand et al.'s (2025) sprint bioenergetic model.
}
\examples{
# Extract Bolt's 100 m data from Graubner and Nixdorf data set.
bolt_100m <- graubner_nixdorf_sprints |>
  dplyr::filter(
    athlete == "Bolt",
    event == "Men's 100 m"
  )

# Compute sprint motion data

bolt_100m_motion_data <- sprint_motion_model_data(
  mean_velocity_splits = bolt_100m$velocity,
  time_splits = bolt_100m$splits,
  distance = bolt_100m$distance,
  reaction_time = bolt_100m$reaction_time[1],
  maximal_velocity = bolt_100m$maximal_velocity[1]
)


# compute the power estimated from the sprint bioenergetic model

bolt_modeled_data <- sprint_bioenergetic_model_data(bolt_100m_motion_data)

percent_error <- sprint_modeled_distance_percentage_error(bolt_modeled_data)

round(percent_error, 2)

}
