% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sprint_motion_metrics.R
\name{sprint_motion_model_data}
\alias{sprint_motion_model_data}
\title{Sprint motion model data (motion metrics, cost of running and metbolic power)}
\usage{
sprint_motion_model_data(
  mean_velocity_splits,
  time_splits,
  distance,
  reaction_time,
  maximal_velocity = NA,
  dt = 0.01
)
}
\arguments{
\item{mean_velocity_splits}{A vector with mean velocity splits over each distance interval (m/s)}

\item{time_splits}{A vector with time splits over each distance interval (s)}

\item{distance}{A vector with the distances at which time splits were measured (m)}

\item{reaction_time}{A double with the reaction time measured on the starting blocks (s)}

\item{maximal_velocity}{A double representing the maximal velocity (in m/s)}

\item{dt}{Time step of the model. Default is set at 0.01 s}
}
\value{
A tibble with the following columns: time, velocity, acceleration, distance, cost of running and power
}
\description{
Compute the velocity, acceleration, distance, cost of running
and power for a sprint event. Based on available time splits and velocity data
provided in the graubner_nixdorf_sprints data set.
}
\examples{
sprint_motion_model_data(mean_velocity_splits = c(0, 5.77, 9.99),
                            time_splits = c(0, 1.88, 2.88),
                            distance = c(0, 10, 20),
                            reaction_time = 0.146,
                            maximal_velocity = 12.34)


}
