% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_performance_simulation.R
\name{sprint_race_time_error_function}
\alias{sprint_race_time_error_function}
\title{Sprint Simulation Race Time Error Function}
\usage{
sprint_race_time_error_function(
  time_performance,
  alactic_capacity,
  lactic_capacity,
  maximal_aerobic_power = 24.5,
  dt = 0.01
)
}
\arguments{
\item{time_performance}{A numeric value representing the running time performance of the sprint (in s). Typically, this performance running time corresponds to the difference between time performance and reaction time (in s).}

\item{alactic_capacity}{A numeric value representing the alactic capacity of the athlete (in J/kg).}

\item{lactic_capacity}{A numeric value representing the lactic capacity of the athlete (in J/kg).}

\item{maximal_aerobic_power}{A numeric value representing the maximal aerobic power of the athlete (in W/kg). Default is set at 24.5 W/kg.}

\item{dt}{A numeric value representing the time step for the simulation (in s). Default is set at 0.01 s.}
}
\value{
A numeric value representing the distance traveled during the sprint (in m).
}
\description{
Error function that computes the distance traveled over a sprint of a given duration. A simulation is made
using the sprint bioenergetic models well as alactic and lactic capacities as well as maximal aerobic power.
The function can be used to find the running time corresponding to classical sprint running distances (e.g. 100 m, 200 m, etc.)
}
\examples{

# Take Bolt 100 m

total_time <- 9.58
reaction_time <- 0.146

performance_running_time <- total_time - reaction_time

# use Briand et al. (2025) alactic and lactic capacities estimates for the simulation

alactic_capacity <- 328
lactic_capacity <- 1460

# apply the error function. We expect the output to be close to 100 m

sprint_race_time_error_function(performance_running_time, alactic_capacity, lactic_capacity)


}
