% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_extract_capacities_from_athlete_perf.R
\name{sprint_simulation_metrics}
\alias{sprint_simulation_metrics}
\title{Sprint Simulation Metrics}
\usage{
sprint_simulation_metrics(
  time_performance,
  distance,
  alactic_capacity = 328,
  lactic_capacity = 1460,
  maximal_aerobic_power = 24.5,
  dt = 0.01
)
}
\arguments{
\item{time_performance}{A numeric value representing the time performance of the sprint (in s).}

\item{distance}{A numeric value representing the distance of the sprint (in m).}

\item{alactic_capacity}{A numeric value representing the alactic capacity of the athlete (in J/kg). Default is set at 328 J/kg.}

\item{lactic_capacity}{A numeric value representing the lactic capacity of the athlete (in J/kg). Default is set at 1460 J/kg.}

\item{maximal_aerobic_power}{A numeric value representing the maximal aerobic power of the athlete (in W/kg). Default is set at 24.5 W/kg.}

\item{dt}{A numeric value representing the time step for the simulation (in s). Default is set at 0.01 s.}
}
\value{
A tibble containing the following columns:
\itemize{
\item time_performance: The time performance of the sprint (in s).
\item distance: The actual distance of the sprint (in m).
\item distance_estimated: The estimated distance traveled during the sprint (in m).
\item distance_error: The percentage error between the estimated and actual distance traveled during the sprint.
\item total_energy: The total energy expenditure during the sprint (in J/kg).
\item alactic_energy: The energy expenditure from the alactic pathway (in J/kg).
\item lactic_energy: The energy expenditure from the lactic pathway (in J/kg).
\item aerobic_energy: The energy expenditure from the aerobic pathway (in J/kg).
}
}
\description{
Provides metrics synthesizing a sprint simulation of a given time performance for a set running distance.
The simulation is based on an athlete's alactic and lactic capacities, as well as maximal aerobic power.
The returned metrics include the estimated distance traveled, the error with respect to the actual distance, total energy expenditure,
as well as the energy expenditure from the three metabolic pathways.
}
\examples{

# Let's take Bolt 100 m performance

time_bolt <- 9.58
reaction_time <- 0.146

running_time_performance <- time_bolt - reaction_time

distance <- 100

# extract the metrics

sprint_simulation_metrics(running_time_performance, distance)


}
