% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_extract_capacities_from_athlete_perf.R
\name{sprint_simulation_metrics_briand_article}
\alias{sprint_simulation_metrics_briand_article}
\title{Sprint Simulation Metrics Table (as in Briand et al. (2025) article)}
\usage{
sprint_simulation_metrics_briand_article(
  male_alactic_capacity = 328,
  male_lactic_capacity = 1460,
  female_alactic_capacity = 218,
  female_lactic_capacity = 1295,
  male_maximal_aerobic_power = 24.5,
  female_maximal_aerobic_power = 21,
  dt = 0.01
)
}
\arguments{
\item{male_alactic_capacity}{A numeric value representing the male athlete's alactic capacity (in J/kg). Default is set at 328 J/kg.}

\item{male_lactic_capacity}{A numeric value representing the male athlete's lactic capacity (in J/kg). Default is set at 1460 J/kg.}

\item{female_alactic_capacity}{A numeric value representing the female athlete's lactic capacity (in J/kg). Default is set at 218 J/kg.}

\item{female_lactic_capacity}{A numeric value representing the female athlete's lactic capacity (in J/kg). Default is set at 1295 J/kg.}

\item{male_maximal_aerobic_power}{A numeric value representing the male athlete's maximal aerobic power (in W/kg). Default is set at 24.5 W/kg.}

\item{female_maximal_aerobic_power}{A numeric value representing the female athlete's maximal aerobic power (in W/kg). Default is set at 21 W/kg.}

\item{dt}{A numeric value representing the time step for the simulation (in s). Default is set at 0.01 s.}
}
\value{
A tinytable object containing the following columns:
\itemize{
\item Event: The name of the event.
\item Time Performance: The time performance of the sprint (in s).
\item Actual Distance: The actual distance of the sprint (in m).
\item Estimated Distance: The estimated distance traveled during the sprint (in m).
\item Distance Error: The percentage error between the estimated and actual distance traveled during the sprint.
\item Total Energy: The total energy expenditure during the sprint (in J/kg).
\item Alactic Energy: The energy expenditure from the alactic pathway (in J/kg).
\item Lactic Energy: The energy expenditure from the lactic pathway (in J/kg).
\item Aerobic Energy: The energy expenditure from the aerobic pathway (in J/kg).
}
}
\description{
Provides a table of metrics synthesizing a sprint simulation of fictitious male and female athletes able to
realise respectively the men's and women's 100, 200 and 400 performances from the \code{graubner_nixdorf_sprints} dataset.
The simulation is based on an athlete's alactic and lactic capacities, as well as maximal aerobic power.
The returned metrics include the estimated distance traveled, the error with respect to the actual distance, total energy expenditure,
as well as the energy expenditure from the three metabolic pathways for each sprint event. The table provide similar information as the one presented in
Briand et al. (2025) article presenting the sprint bioenergetic model.
}
\examples{
sprint_simulation_metrics_briand_article()
}
