% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_performance_simulation.R
\name{sprint_time_perf_simulation}
\alias{sprint_time_perf_simulation}
\title{Sprint Time Performance Simulation}
\usage{
sprint_time_perf_simulation(
  target_distances,
  alactic_capacity,
  lactic_capacity,
  maximal_aerobic_power = 24.5,
  dt = 0.01,
  reaction_time = 0.15
)
}
\arguments{
\item{target_distances}{A numeric vector representing the target distances for the sprint simulation (in m).}

\item{alactic_capacity}{A numeric value representing the alactic capacity of the athlete (in J/kg).}

\item{lactic_capacity}{A numeric value representing the lactic capacity of the athlete (in J/kg).}

\item{maximal_aerobic_power}{A numeric value representing the maximal aerobic power of the athlete (in W/kg). Default is set at 24.5 W/kg.}

\item{dt}{A numeric value representing the time step for the simulation (in s). Default is set at 0.01 s.}

\item{reaction_time}{A numeric value representing the reaction time of the athlete (in s). Default is set at 0.15 s.}
}
\value{
A tibble with the following columns:
\itemize{
\item Distance: The target distance for the sprint (in m).
\item Optimal_Time: The running time for the sprint, without considering reaction time (in s).
\item Actual_Time: The running time for the sprint, including reaction time (in s).
}
}
\description{
Simulates the time performances on target distance given an athlete's estimated alactic and lactic capacities as well as maximal aerobic power.
The simulation is based on the sprint bioenergetic model. The returned results present the \code{optimal_time}, which corresponds to running time, as well as \code{actual_time},
which corresponds to the running time plus reaction time. The reaction time can be provided as a function parameter. The simulation can be used to estimate the running time of a sprint on classical sprint distances (e.g. 100 m, 200 m, etc.).
}
\examples{

target_distances <- c(60, 100, 200, 300, 400)
 # Use Briand et al. (2025) alactic and lactic capacities estimates for the simulation
 alactic_capacity <- 328
 lactic_capacity <- 1460

# apply the simulation

sprint_time_perf_simulation(target_distances, alactic_capacity, lactic_capacity)

}
