% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_motion_models_diagnostics.R
\name{sprint_velocity_least_squares_model}
\alias{sprint_velocity_least_squares_model}
\title{Sprint Acceleration Velocity Model (Non-Linear Least-Squares Approach)}
\usage{
sprint_velocity_least_squares_model(time, velocity, reaction_time)
}
\arguments{
\item{time}{A vector of time points (in s)}

\item{velocity}{A vector containing average velocity over each interval (in m/s)}

\item{reaction_time}{A double corresponding to the reaction time (in s)}
}
\value{
A fitted nlsLM model object.
}
\description{
Fits a non-linear least-squares model to the sprint acceleration velocity data.
}
\examples{

men_200 <- graubner_nixdorf_sprints |>
  dplyr::filter(event == "Men's 200 m")

time <- find_time_velocity(men_200$splits, men_200$reaction_time[1])
velocity <- men_200$velocity
reaction_time <- men_200$reaction_time[1]

# generate the model
sprint_velocity_least_squares_model(time = time, velocity = velocity, reaction_time = reaction_time)

}
