% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_motion_computations.R
\name{velocity_sprint_model}
\alias{velocity_sprint_model}
\title{Instantaneous Velocity During Sprint Running (Acceleration and Deceleration phases)}
\usage{
velocity_sprint_model(
  time,
  time_maximal_velocity,
  maximal_velocity,
  tau,
  fitted_maximal_velocity,
  decel_rate
)
}
\arguments{
\item{time}{A vector of time points (in s)}

\item{time_maximal_velocity}{A double representing the time at which maximal velocity is reached (in s)}

\item{maximal_velocity}{A double representing the maximal velocity (in m/s)}

\item{tau}{A double representing the time constant of the exponential rise in velocity (in s)}

\item{fitted_maximal_velocity}{A double representing the maximal velocity fitted using the model for instantaneous velocity in acceleration phase (in m/s)}

\item{decel_rate}{A double representing the rate of deceleration (in m/s^2)}
}
\value{
A vector of instantaneous velocities (in m/s)
}
\description{
Estimates the instantaneous velocity during the acceleration and deceleration phases of a sprint run.
The calculation is based on an exponential rise in velocity during the acceleration phase and a constant rate of deceleration during the deceleration phase.
}
\examples{
velocity_sprint_model(1:10, 5, 10, 1.5, 10, 0.1)

}
