% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_integrals.R
\name{alactic_energy_percentage}
\alias{alactic_energy_percentage}
\title{Alactic Energy Percentage}
\usage{
alactic_energy_percentage(data)
}
\arguments{
\item{data}{A tibble with a time series of alactic power values and associated times. The tibble must minimally have the following columns: time (s), power (W/kg), power_alactic (W/kg)}
}
\value{
A double representing the percentage (\%) of anaerobic alactic energy expenditure relative to total energy expenditure.
}
\description{
Computes the percentage of the total energy expenditure that is attributed to anaerobic alactic energy pathway.
}
\examples{

data <- tibble::tibble(
time = c(0, 1, 2, 3, 4),
power = c(0, 1, 2, 1, 0),
power_alactic = c(0, 0.5, 1, 0.5, 0))

alactic_energy_percentage(data)


}
