% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_of_running.R
\name{cost_running}
\alias{cost_running}
\title{Cost of Running}
\usage{
cost_running(acceleration, velocity)
}
\arguments{
\item{acceleration}{a vector of accelerations (m/s^2)}

\item{velocity}{a vector of velocities (m/s)}
}
\value{
a vector of costs of running (J/kg/m)
}
\description{
Computes the instantaneous cost of running based on velocity and acceleration data,
based on the method of di prampero et al. (2005, 2018) and as described in Briand et al. (2025).
The function considers scenarios where the deceleration is initiated by the athlete
changing the position of the center of mass. The function can be applied for example
on gps data derived from athletes performing changes of directions in a team sport
as in di Prampero and Osgnach (2018).
}
\examples{
cost_running(acceleration = c(2, 4, 5), velocity = c(2,4,5))

}
