% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_motion_models_diagostics.R
\name{diagnose_velocity_model_acc}
\alias{diagnose_velocity_model_acc}
\title{Test the Assumptions of the Least Squares Acceleration Velocity Model}
\usage{
diagnose_velocity_model_acc(
  mean_velocity_splits,
  time_splits,
  distance,
  reaction_time,
  sprint_total_distance = 100,
  histogram_binwidth = 0.1
)
}
\arguments{
\item{mean_velocity_splits}{A vector with mean velocity splits over each distance interval (m/s)}

\item{time_splits}{A vector with time splits over each distance interval (s)}

\item{distance}{A vector with the distances at which time splits were measured (m)}

\item{reaction_time}{A double with the reaction time measured on the starting blocks (s)}

\item{sprint_total_distance}{A numeric value representing the distance (in m) (default is 100 m)}

\item{histogram_binwidth}{A numeric value representing the binwidth for the histogram of residuals (default is 0.1)}
}
\value{
A list containing the RMSE, RSE, and a list of ggplot objects for each diagnostic plot.
}
\description{
Tests the assumptions of the least squares acceleration velocity model by plotting observed vs. fitted values,
residuals vs. fitted values, Q–Q plot of residuals, and histogram of residuals.
}
\examples{

men_200 <- graubner_nixdorf_sprints |>
  dplyr::filter(event == "Men's 200 m")

diagnose_velocity_model_acc(
  mean_velocity_splits = men_200$velocity,
  time_splits = men_200$splits,
  distance = men_200$distance,
  reaction_time = men_200$reaction_time[1],
  sprint_total_distance = 200
)

}
