% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_integrals.R
\name{energy_total}
\alias{energy_total}
\title{Total Energy Expenditure}
\usage{
energy_total(data)
}
\arguments{
\item{data}{A tibble with a time series of power values and associated times. The tibble must minimally have the following columns: time (s), power (W/kg)}
}
\value{
A double with the total energy expended (J/kg)
}
\description{
Computes the integral of the total metabolic power over time to obtain total energy expenditure.
}
\examples{
data <- tibble::tibble(
 time = c(0, 1, 2, 3, 4),
 power = c(0, 1, 2, 1, 0)
)

energy_total(data)



}
