% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_alactic_energy_duration.R
\name{find_max_al}
\alias{find_max_al}
\title{Find Sprint Maximal Alactic Power}
\usage{
find_max_al(
  duration,
  alactic_capacity,
  mu = -0.4,
  sigma = 1,
  mu_al = 1.75,
  sigma_al = 1.5
)
}
\arguments{
\item{duration}{A numeric value representing the sprint duration (in seconds).}

\item{alactic_capacity}{A numeric value representing the lactic capacity (in J/kg).}

\item{mu}{A double. Parameter setting the peak of the log-normal distribution.Default is -0.4}

\item{sigma}{A double. Parameter setting the decay of the log-normal distribution. Default is 1}

\item{mu_al}{A double representing the peak of the log-normal distribution. Default is 1.75.}

\item{sigma_al}{A double representing the decay of the log-normal distribution. Default is 1.5.}
}
\value{
A numeric value representing the maximal lactic power (in W/kg) that can be sustained over the given sprint duration.
}
\description{
Knowing an athlete maximal alactic capacity, the function finds
the maximal alactic power that can be sustained over a given sprint duration
}
\examples{

duration <- 20
alactic_capacity <- 350

find_max_al(duration, alactic_capacity)

}
