% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_model_formalisation.R
\name{fit_lactic_gof_metrics}
\alias{fit_lactic_gof_metrics}
\title{Goodness of Fit (Gof) of Lactic Power Model (Sprint)}
\usage{
fit_lactic_gof_metrics(data, k1 = 2.75, k2 = 35)
}
\arguments{
\item{data}{A tibble containing approximate power distributions over the course of the sprint with at least the following columns: time (s), power_lactic (W/kg)}

\item{k1}{A double. Time constant of the first rising exponential (s).}

\item{k2}{A double. Time constant of the second decaying exponential (s).}
}
\value{
A list with the following elements: r_squared, residual_se, AIC, BIC
}
\description{
Computes the goodness of fit metrics for the formalized sprint lactic power model (with k1 set to 2.75 s, and k2 set to 35 s)
}
\examples{

# Extract the data for the 100 m
men_100 <- graubner_nixdorf_sprints |>
  dplyr::filter(event == "Men's 100 m")


# Get the sprint motion data for both men and women

sprint_data <- sprint_motion_model_data(
  mean_velocity_splits = men_100$velocity,
  time_splits = men_100$splits,
  distance = men_100$distance,
  reaction_time = men_100$reaction_time[1],
  maximal_velocity = men_100$maximal_velocity[1]
)

sprint_approx_power_distributions <- sprint_approx_power_distributions(sprint_data,
  maximal_aerobic_power = 24.5,
  basal_metabolic_rate = 1.2
)

# Fit the bi-exponential distribution to the lactic power distribution

fit_lactic_gof_metrics(sprint_approx_power_distributions)

}
