% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sprint_lactic_duration_model.R
\name{get_lactic_model_data}
\alias{get_lactic_model_data}
\title{Generate model predictions for the sprint lactic energy-duration model}
\usage{
get_lactic_model_data(
  lactic_energy_duration,
  sex_label,
  t1 = 20,
  t2 = 1500,
  max_duration = 100
)
}
\arguments{
\item{lactic_energy_duration}{A tibble with the following columns: duration (s), lactic_energy (J/kg), corresponding to the average lactic energy expenditure over running duration. This data is used to fit the lactic energy model.}

\item{sex_label}{A string label (e.g., "male", "Female") used for grouping the output.}

\item{t1}{A numeric value representing the time constant for the first exponential decay (default is 20 seconds).}

\item{t2}{A numeric value representing the time constant for the second exponential decay (default is 1500 seconds).}

\item{max_duration}{Maximum duration for which to compute the model (default = 100 s).}
}
\value{
A tibble with \code{duration}, \code{energy}, and \code{sex} columns representing the predicted model values.
}
\description{
Computes the modeled lactic energy output across a sequence of durations,
based on the estimated lactic capacity from empirical data and bi-exponential lactic energy-duration model.
}
\examples{
data <- sprint_lactic_energy_duration_graubner_nixdorf()
get_lactic_model_data(data, sex_label = "male")
}
