% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sprint_alactic_duration_model.R
\name{plot_sprint_alactic_duration_briand_article}
\alias{plot_sprint_alactic_duration_briand_article}
\title{Plot alactic power-duration model fits for male and female sprinters}
\usage{
plot_sprint_alactic_duration_briand_article(
  data = graubner_nixdorf_sprints,
  mu_al = 1.75,
  sigma_al = 1.5,
  line_color = "purple",
  point_color = "darkorchid4"
)
}
\arguments{
\item{data}{A data frame containing sprint alactic power estimations (default: \code{graubner_nixdorf_sprints}).}

\item{mu_al}{A double representing the peak of the log-normal distribution. Default is 1.75.}

\item{sigma_al}{A double representing the decay of the log-normal distribution. Default is 1.5.}

\item{line_color}{Color for model line (default: "purple").}

\item{point_color}{Color for data points (default: "darkorchid4").}
}
\value{
A ggplot object with overlaid male and female model fits and alactic power over different running durations.
}
\description{
Provides observed alactic power values and model predictions
for male and female athletes using the log-normal decay model.The function using the graubner_nixdorf_sprints data set
generates the same Figure as the one presented in Briand et al. 2025.
}
\examples{
plot_sprint_alactic_duration_briand_article()
}
