% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sprint_lactic_duration_model.R
\name{plot_sprint_lactic_duration_briand_article}
\alias{plot_sprint_lactic_duration_briand_article}
\title{Plot lactic energy-duration model fits for male and female sprinters}
\usage{
plot_sprint_lactic_duration_briand_article(
  data = graubner_nixdorf_sprints,
  t1 = 20,
  t2 = 1500,
  line_color = "#457B9D",
  point_color = "darkblue",
  mu = -0.4,
  sigma = 1,
  k1 = 2.75,
  k2 = 35,
  dt = 0.01,
  cost_running_flat = 3.8,
  slope_equation = "original"
)
}
\arguments{
\item{data}{A data frame containing sprint lactic energy estimations (default: \code{graubner_nixdorf_sprints}).}

\item{t1}{A numeric value representing the time constant for the first exponential decay (default is 20 seconds).}

\item{t2}{A numeric value representing the time constant for the second exponential decay (default is 1500 seconds).}

\item{line_color}{Color for model line (default: "#457B9D").}

\item{point_color}{Color for data points (default: "darkblue").}

\item{mu}{A double. Parameter setting the peak of the log-normal distribution.Default is -0.4}

\item{sigma}{A double. Parameter setting the decay of the log-normal distribution. Default is 1}

\item{k1}{A double. Time constant of the first rising exponential (s). Default is 2.75}

\item{k2}{A double. Time constant of the second decaying exponential (s). Default is 35}

\item{cost_running_flat}{a numeric value representing the cost of running on a flat surface (default for this function is 3.8 J/kg/m, as used in Briand et al. 2025)}

\item{slope_equation}{a string indicating the slope equation to use, either "original" or "extended". The "original" equation is based on the equation relating running at constant speed on different slopes porposed by Minetti et al. (2002), while the "extended" equation is based on similar equation for extended slopes proposed by Minetti and Pavei (2018).}
}
\value{
A ggplot object with overlaid male and female model fits and lactic energy over different running durations.
}
\description{
Provides observed lactic energy values and model predictions
for male and female athletes using the bi-exponential lactic energy-duration model.The function using the \code{graubner_nixdorf_sprints} data set
generates the same Figure as the one presented in Briand et al. 2025.
}
\examples{
plot_sprint_lactic_duration_briand_article()
}
