% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_model_formalisation.R
\name{sprint_alactic_model_gof_comp}
\alias{sprint_alactic_model_gof_comp}
\title{Table of Comparison of the Optimal and Formalized/Simplified Alactic Model Parameters and Goodness of Fit Metrics (On Berlin 2009 Sprint Performances)}
\usage{
sprint_alactic_model_gof_comp(data = graubner_nixdorf_sprints)
}
\arguments{
\item{data}{A tibble with the following: distance (m), splits (s), velocity (m/s), reaction_time (s), maximal_velocity (m/s) and event (character). Default is Graubner and Nixdorf (2009) sprint data.}
}
\value{
A tibble comparing the optimal and formalized/simplified alactic model fitted parameters and goodness of fit metrics over different events: event (character), pal_max (W/kg), mu, sigma, AIC, BIC and R-squared.
}
\description{
Provides a table comparing the optimal and formalized/simplified alactic model fitted parameters and goodness of fit metrics over different events. The optimal model
is a model where the maximal alactic power as well as mu and sigma parameters of the log-normal distribution are fitted to the data.
The formalized model is a model where the mu and sigma parameters of the log-normal distribution are set at -0.4 and 1 respectively.
}
\examples{

sprint_alactic_model_gof_comp()

}
