% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_model_formalisation.R
\name{sprint_alactic_model_gof_metrics}
\alias{sprint_alactic_model_gof_metrics}
\title{Goodness of Fit Metrics for the formalised Alactic Model (On Berlin 2009 Sprint Performances)}
\usage{
sprint_alactic_model_gof_metrics(data = graubner_nixdorf_sprints)
}
\arguments{
\item{data}{A tibble with the following: distance (m), splits (s), velocity (m/s), reaction_time (s), maximal_velocity (m/s) and event (character). Default is Graubner and Nixdorf (2009) sprint data.}
}
\value{
A tibble with the goodness of fit metrics for the formalized/simplified alactic model: event (character), AIC, BIC, residual squared error and R-squared.
}
\description{
Computes the goodness of fit metrics (AIC, BIC and R-squared) for the alactic model based on the sprint data from Graubner and Nixdorf (2009) for the men's
and women's 100, 200 and 400 m.
}
\examples{


sprint_alactic_model_gof_metrics()

}
