% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_model_formalisation.R
\name{sprint_alactic_model_optimal_params}
\alias{sprint_alactic_model_optimal_params}
\title{Optimal Parameters for the Alactic Model (On Berlin 2009 Sprint Performances)}
\usage{
sprint_alactic_model_optimal_params(data = graubner_nixdorf_sprints)
}
\arguments{
\item{data}{A tibble with the following: distance (m), splits (s), velocity (m/s), reaction_time (s), maximal_velocity (m/s) and event (character). Default is Graubner and Nixdorf (2009) sprint data.}
}
\value{
A tibble with the optimal parameters for the alactic model: event (character), pal_max (W/kg), mu and sigma.
}
\description{
Computes the optimal parameters for the alactic model based on the sprint data from Graubner and Nixdorf (2009) for the men's
and women's 100, 200 and 400 m.
}
\examples{

sprint_alactic_model_optimal_params()

}
