% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_lactic_energy_duration.R
\name{sprint_lactic_capacity}
\alias{sprint_lactic_capacity}
\title{Lactic Capacity from Lactic Energy Duration Data}
\usage{
sprint_lactic_capacity(lactic_energy_duration, t1 = 20, t2 = 1500)
}
\arguments{
\item{lactic_energy_duration}{A tibble with the following columns: duration (s), lactic_energy (J/kg), corresponding to the average lactic energy expenditure over running duration. This data is used to fit the lactic energy model.}

\item{t1}{A numeric value representing the time constant for the first exponential decay (default is 20 seconds).}

\item{t2}{A numeric value representing the time constant for the second exponential decay (default is 1500 seconds).}
}
\value{
A numeric value representing the lactic capacity (in J/kg) extracted from the fitted lactic energy model.
}
\description{
Extracts the lactic capacity from the fitted lactic energy model based on the lactic energy duration data.
}
\examples{
men_lactic_energy <- sprint_lactic_energy_duration_graubner_nixdorf()
men_lactic_energy

 # Extract lactic capacity from the data

 sprint_lactic_capacity(men_lactic_energy)

}
