% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_lactic_energy_duration.R
\name{sprint_lactic_duration_model_fit}
\alias{sprint_lactic_duration_model_fit}
\title{Sprint Lactic Duration Model Fit}
\usage{
sprint_lactic_duration_model_fit(lactic_energy_duration, t1 = 20, t2 = 1500)
}
\arguments{
\item{lactic_energy_duration}{A tibble with the following columns: duration (s), lactic_energy (J/kg), corresponding to the average lactic energy expenditure over running duration. This data is used to fit the lactic energy model.}

\item{t1}{A numeric value representing the time constant for the first exponential decay (default is 20 seconds).}

\item{t2}{A numeric value representing the time constant for the second exponential decay (default is 1500 seconds).}
}
\value{
A fitted model object of class "nls" representing the lactic energy model fitted to the lactic energy duration data.
}
\description{
Performs a non-linear least squares fitting of the lactic energy
model to the lactic energy over running duration data.
}
\examples{
men_lactic_energy <- sprint_lactic_energy_duration_graubner_nixdorf()
men_lactic_energy

 # Fit lactic model to energy

 sprint_lactic_duration_model_fit(men_lactic_energy)

}
