% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_performance_simulation.R
\name{sprint_race_time_error_function}
\alias{sprint_race_time_error_function}
\title{Sprint Simulation Race Time Error Function}
\usage{
sprint_race_time_error_function(
  time_performance,
  alactic_capacity,
  lactic_capacity,
  maximal_aerobic_power = 24.5,
  mu = -0.4,
  sigma = 1,
  k1 = 2.75,
  k2 = 35,
  dt = 0.01,
  cost_running_flat = 3.8,
  slope_equation = "original"
)
}
\arguments{
\item{time_performance}{A numeric value representing the running time performance of the sprint (in s). Typically, this performance running time corresponds to the difference between time performance and reaction time (in s).}

\item{alactic_capacity}{A numeric value representing the alactic capacity of the athlete (in J/kg).}

\item{lactic_capacity}{A numeric value representing the lactic capacity of the athlete (in J/kg).}

\item{maximal_aerobic_power}{A numeric value representing the maximal aerobic power of the athlete (in W/kg). Default is set at 24.5 W/kg.}

\item{mu}{A double. Parameter setting the peak of the log-normal distribution.Default is -0.4}

\item{sigma}{A double. Parameter setting the decay of the log-normal distribution. Default is 1}

\item{k1}{A double. Time constant of the first rising exponential (s). Default is 2.75}

\item{k2}{A double. Time constant of the second decaying exponential (s). Default is 35}

\item{dt}{a numeric value representing the time step at which power is provided (in seconds). Default is 0.01 seconds.}

\item{cost_running_flat}{a numeric value representing the cost of running on a flat surface (default for this function is 3.8 J/kg/m, as used in Briand et al. 2025)}

\item{slope_equation}{a string indicating the slope equation to use, either "original" or "extended". The "original" equation is based on the equation relating running at constant speed on different slopes porposed by Minetti et al. (2002), while the "extended" equation is based on similar equation for extended slopes proposed by Minetti and Pavei (2018).}
}
\value{
A numeric value representing the distance traveled during the sprint (in m).
}
\description{
Error function that computes the distance traveled over a sprint of a given duration. A simulation is made
using the sprint bioenergetic models well as alactic and lactic capacities as well as maximal aerobic power.
The function can be used to find the running time corresponding to classical sprint running distances (e.g. 100 m, 200 m, etc.)
}
\examples{

# Take Bolt 100 m

total_time <- 9.58
reaction_time <- 0.146

performance_running_time <- total_time - reaction_time

# use Briand et al. (2025) alactic and lactic capacities estimates for the simulation

alactic_capacity <- 328
lactic_capacity <- 1460

# apply the error function. We expect the output to be close to 100 m

sprint_race_time_error_function(performance_running_time, alactic_capacity, lactic_capacity)


}
