% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_extract_capacities_from_athlete_perf.R
\name{sprint_simulation_metrics}
\alias{sprint_simulation_metrics}
\title{Sprint Simulation Metrics}
\usage{
sprint_simulation_metrics(
  time_performance,
  distance,
  alactic_capacity = 328,
  lactic_capacity = 1460,
  maximal_aerobic_power = 24.5,
  dt = 0.01,
  mu = -0.4,
  sigma = 1,
  k1 = 2.75,
  k2 = 35,
  cost_running_flat = 3.6,
  slope_equation = "original"
)
}
\arguments{
\item{time_performance}{A numeric value representing the time performance of the sprint (in s).}

\item{distance}{A numeric value representing the distance of the sprint (in m).}

\item{alactic_capacity}{A numeric value representing the alactic capacity of the athlete (in J/kg). Default is set at 328 J/kg.}

\item{lactic_capacity}{A numeric value representing the lactic capacity of the athlete (in J/kg). Default is set at 1460 J/kg.}

\item{maximal_aerobic_power}{A double. Maximal aerobic power (W/kg). Default is 24.5}

\item{dt}{a numeric value representing the time step at which power is provided (in seconds). Default is 0.01 seconds.}

\item{mu}{A double. Parameter setting the peak of the log-normal distribution.Default is -0.4}

\item{sigma}{A double. Parameter setting the decay of the log-normal distribution. Default is 1}

\item{k1}{A double. Time constant of the first rising exponential (s). Default is 2.75}

\item{k2}{A double. Time constant of the second decaying exponential (s). Default is 35}

\item{cost_running_flat}{a numeric value representing the cost of running on a flat surface (default is 3.6 J/kg/m, as suggested in Minetti et al. 2002)}

\item{slope_equation}{a string indicating the slope equation to use, either "original" or "extended". The "original" equation is based on the equation relating running at constant speed on different slopes porposed by Minetti et al. (2002), while the "extended" equation is based on similar equation for extended slopes proposed by Minetti and Pavei (2018).}
}
\value{
A tibble containing the following columns:
\itemize{
\item time_performance: The time performance of the sprint (in s).
\item distance: The actual distance of the sprint (in m).
\item distance_estimated: The estimated distance traveled during the sprint (in m).
\item distance_error: The percentage error between the estimated and actual distance traveled during the sprint.
\item total_energy: The total energy expenditure during the sprint (in J/kg).
\item alactic_energy: The energy expenditure from the alactic pathway (in J/kg).
\item lactic_energy: The energy expenditure from the lactic pathway (in J/kg).
\item aerobic_energy: The energy expenditure from the aerobic pathway (in J/kg).
}
}
\description{
Provides metrics synthesizing a sprint simulation of a given time performance for a set running distance.
The simulation is based on an athlete's alactic and lactic capacities, as well as maximal aerobic power.
The returned metrics include the estimated distance traveled, the error with respect to the actual distance, total energy expenditure,
as well as the energy expenditure from the three metabolic pathways.
}
\examples{

# Let's take Bolt 100 m performance

time_bolt <- 9.58
reaction_time <- 0.146

running_time_performance <- time_bolt - reaction_time

distance <- 100

# extract the metrics

sprint_simulation_metrics(running_time_performance, distance)


}
