% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_performance_simulation.R
\name{sprint_time_perf_simulation_briand_table}
\alias{sprint_time_perf_simulation_briand_table}
\title{Sprint Time Performance Simulation Formatted Table}
\usage{
sprint_time_perf_simulation_briand_table(
  target_distances,
  alactic_capacity,
  lactic_capacity,
  maximal_aerobic_power = 24.5,
  mu = -0.4,
  sigma = 1,
  k1 = 2.75,
  k2 = 35,
  dt = 0.01,
  cost_running_flat = 3.8,
  slope_equation = "original",
  reaction_time = 0.15
)
}
\arguments{
\item{target_distances}{A numeric vector representing the target distances for the sprint simulation (in m).}

\item{alactic_capacity}{A numeric value representing the alactic capacity of the athlete (in J/kg).}

\item{lactic_capacity}{A numeric value representing the lactic capacity of the athlete (in J/kg).}

\item{maximal_aerobic_power}{A double. Maximal aerobic power (W/kg). Default is 24.5}

\item{mu}{A double. Parameter setting the peak of the log-normal distribution.Default is -0.4}

\item{sigma}{A double. Parameter setting the decay of the log-normal distribution. Default is 1}

\item{k1}{A double. Time constant of the first rising exponential (s). Default is 2.75}

\item{k2}{A double. Time constant of the second decaying exponential (s). Default is 35}

\item{dt}{a numeric value representing the time step at which power is provided (in seconds). Default is 0.01 seconds.}

\item{cost_running_flat}{a numeric value representing the cost of running on a flat surface (default for this function is 3.8 J/kg/m, as used in Briand et al. 2025)}

\item{slope_equation}{a string indicating the slope equation to use, either "original" or "extended". The "original" equation is based on the equation relating running at constant speed on different slopes porposed by Minetti et al. (2002), while the "extended" equation is based on similar equation for extended slopes proposed by Minetti and Pavei (2018).}

\item{reaction_time}{A numeric value representing the reaction time of the athlete (in s). Default is set at 0.15 s.}
}
\value{
A tinytable with the following columns:
\itemize{
\item Distance: The target distance for the sprint (in m).
\item Running Time: The running time for the sprint, without considering reaction time (in s).
\item Performance Time: The running time for the sprint, including reaction time (in s).
}
}
\description{
Provides a formatted table of simulated time performances on target distance given an athlete's estimated alactic and lactic capacities as well as maximal aerobic power.
The simulation is based on the sprint bioenergetic model. The returned results present the \verb{Running Time}, which corresponds to running time in s, as well as \verb{Performance Time},
which corresponds to the running time plus reaction time (in s). The reaction time can be provided as a function parameter. The simulation can be used to estimate the running time of a sprint on classical sprint distances (e.g. 100 m, 200 m, etc.).
}
\examples{

target_distances <- c(60, 100, 200, 300, 400)
 # Use Briand et al. (2025) alactic and lactic capacities estimates for the simulation
 alactic_capacity <- 328
 lactic_capacity <- 1460

# apply the simulation

sprint_time_perf_simulation_briand_table(target_distances, alactic_capacity, lactic_capacity)
}
